### 2023-06-11 2fe4627
Updated BiBTeX file
Updated bibliography file by removing the uncited items.
