+++ paper/paper.bib
@@ -1,182 +1,26 @@
 
 
 
-@ARTICLE{Figueredo:2009dg,
-   author = {Figueredo, A.~J. and Wolf, P. S.~A.},
-   title = {Assortative pairing and life history strategy -- a cross-cultural study},
-   journal = {Human Nature},
-   volume = {20},
-   pages = {317-330},
-   year = {2009},
-   doi={https://doi.org/10.1007/s12110-009-9068-2}
-}
-
-@MISC{Hao:gidmaps:2014,
-  author = {Hao, Z. and AghaKouchak, A. and Nakhjiri, N. and Farahmand, A},
-  year = {2014},
-  title = {Global integrated drought monitoring and prediction system ({GIDMaPS}) data sets},
-  howpublished = {\emph{figshare} \url{http://dx.doi.org/10.6084/m9.figshare.853801}}
-}
 
 
-@inproceedings{cleary1994jacobian,
-  title={Jacobian formulation for a novel 6-DOF parallel manipulator},
-  author={Cleary, Kevin and Uebel, Mark},
-  booktitle={Proceedings of the 1994 IEEE International Conference on Robotics and Automation},
-  pages={2377--2382},
-  year={1994},
-  organization={IEEE}
-}
 
-@article{kim2003new,
-  title={New dimensionally homogeneous Jacobian matrix formulation by three end-effector points for optimal design of parallel manipulators},
-  author={Kim, Sung-Gaun and Ryu, Jeha},
-  journal={IEEE Transactions on Robotics and Automation},
-  volume={19},
-  number={4},
-  pages={731--736},
-  year={2003},
-  publisher={IEEE}
-}
 
-@article{pond2006formulating,
-  title={Formulating Jacobian matrices for the dexterity analysis of parallel manipulators},
-  author={Pond, Geoffrey and Carretero, Juan A},
-  journal={Mechanism and Machine Theory},
-  volume={41},
-  number={12},
-  pages={1505--1519},
-  year={2006},
-  publisher={Elsevier}
-}
 
-@article{liu2010method,
-  title={A method to formulate a dimensionally homogeneous Jacobian of parallel manipulators},
-  author={Liu, Haitao and Huang, Tian and Chetwynd, Derek G},
-  journal={IEEE Transactions on Robotics},
-  volume={27},
-  number={1},
-  pages={150--156},
-  year={2010},
-  publisher={IEEE}
-}
 
-@article{hu2014formulation,
-  title={Formulation of unified Jacobian for serial-parallel manipulators},
-  author={Hu, Bo},
-  journal={Robotics and Computer-Integrated Manufacturing},
-  volume={30},
-  number={5},
-  pages={460--467},
-  year={2014},
-  publisher={Elsevier}
-}
 
-@ARTICLE{56660,  author={Gosselin, C. and Angeles, J.},  journal={IEEE Transactions on Robotics and Automation},   title={Singularity analysis of closed-loop kinematic chains},   year={1990},  volume={6},  number={3},  pages={281-290}}
+
+
+
+
+
+
+
 
[…]
