### 2023-06-12 9239aab
Updated BiBTeX file
Updated bibliography file by adding missing DOIs.
