+++ paper/paper.bib
@@ -16,7 +16,7 @@
 
 
 
-@INPROCEEDINGS{846382,  author={Kim, D. and Chung, W. and Youm, Y.},  booktitle={Proceedings 2000 ICRA. Millennium Conference. IEEE International Conference on Robotics and Automation. Symposia Proceedings (Cat. No.00CH37065)},   title={Analytic Jacobian of in-parallel manipulators},   year={2000},  volume={3},  number={},  pages={2376-2381 vol.3},  abstract={A general formulation to obtain an analytic Jacobian matrix for in-parallel manipulators has not been developed yet, although several authors addressed it. This paper suggests a consistent approach to obtain the instantaneous kinematic relation of in-parallel manipulators, by extending the results of screw theory to the forward and inverse instantaneous relations, and gives a formula that can obtain a reciprocal screw analytically without considering geometric relations.},  keywords={},  ISSN={1050-4729},  month={April},}
+@INPROCEEDINGS{846382,  author={Kim, D. and Chung, W. and Youm, Y.},  booktitle={Proceedings 2000 ICRA. Millennium Conference. IEEE International Conference on Robotics and Automation. Symposia Proceedings (Cat. No.00CH37065)},   title={Analytic Jacobian of in-parallel manipulators},   year={2000},  volume={3},  number={},  pages={2376-2381 vol.3},  abstract={A general formulation to obtain an analytic Jacobian matrix for in-parallel manipulators has not been developed yet, although several authors addressed it. This paper suggests a consistent approach to obtain the instantaneous kinematic relation of in-parallel manipulators, by extending the results of screw theory to the forward and inverse instantaneous relations, and gives a formula that can obtain a reciprocal screw analytically without considering geometric relations.},  keywords={},  ISSN={1050-4729},  month={April}, doi={10.1109/robot.2000.846382},}
 
 
 
@@ -30,7 +30,8 @@
   number={3},
   pages={547--570},
   year={2015},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s10846-014-0024-y},
 }
 
 
@@ -63,7 +64,7 @@
 
 
 
-@ARTICLE{1220722,  author={Sung-Gaun Kim and Ryu, J.},  journal={IEEE Transactions on Robotics and Automation},   title={New dimensionally homogeneous Jacobian matrix formulation by three end-effector points for optimal design of parallel manipulators},   year={2003},  volume={19},  number={4},  pages={731-736}}
+@ARTICLE{1220722,  author={Sung-Gaun Kim and Ryu, J.},  journal={IEEE Transactions on Robotics and Automation},   title={New dimensionally homogeneous Jacobian matrix formulation by three end-effector points for optimal design of parallel manipulators},   year={2003},  volume={19},  number={4},  pages={731-736}, doi={10.1109/tra.2003.814496}}
 
 
 @article{ALTUZARRA20061407,
@@ -76,10 +77,11 @@ year = {2006},
 issn = {0094-114X},
 author = {O. Altuzarra and O. Salgado and V. Petuya and A. Hernández},
 keywords = {Velocity equation, Dimensionless Jacobian, Motion space, Parallel manipulators},
[…]
