### 2023-06-12 #5536 @suneeshjacob
@arfon - i removed uncited references and added missing DOIs.
