### 2023-06-26 c8d6276
Update Mathematics behind Jacobian formulation.md
Made a small fix
