### 2023-07-22 #5536 @suneeshjacob
@gkthiruvathukal a close-to-ideal reviewer (whose JOSS paper is very much related to mine) would be Jan Baumgärtner
`liquidcronos`, but unfortunately i couldn't find them in the list of people that have agreed to review. Here are some people from the list who may be potential reviewers: `destogl`, `gfadini` and `JHartzer`. If anything else is needed, please let me know. Thanks!
