### 2023-08-31 #5536 @suneeshjacob
@gkthiruvathukal

Dear Prof. George,

Greetings.

We are eagerly waiting to hear from the reviewers, as we submitted the paper on 24th May 2023. We would appreciate if you can update us regarding this.

If you need any other reviewers to expedite the process, we would be happy to propose some names of experts.

Kind regards,
Dr. Suneesh & Dr. Rituparna
