### 2023-10-04 #5536 @Kevin-Mattheus-Moerman
@CameronDevine, @bmagyar, @Amudtogal, @sea-bass, @JHartzer, @destogl, @palmieri would you be interested in reviewing this submission about robotics for the [Journal of Open Source Software](https://joss.theoj.org/). The title is: "Automatic Computation for Robot Design (ACRoD): A Python package for numerical calculation of Jacobian of a robot at a given configuration around a specified end-effector point".
Our reviews take place on GitHub and focus on [the software](https://github.com/suneeshjacob/ACRoD) as well as a [short paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05536/joss.05536/10.21105.joss.05536.pdf).

You can let us know here in this "pre-review" issue, if you are interested.

Once we have sufficient reviewers a dedicated review issues is triggered featuring checklists to guide reviewers through the process. Here is an example of a completed review: https://github.com/openjournals/joss-reviews/issues/5670

Let me know if you have any questions.
