### 2023-10-05 #5927 @editorialbot
[REVIEW]: Automatic Computation for Robot Design (ACRoD): A Python package for numerical calculation of Jacobian of a robot at a given configuration around a specified end-effector point
**Submitting author:** @suneeshjacob (<a href="http://orcid.org/0000-0002-9855-9368">Akkarapakam Suneesh Jacob</a>)
**Repository:** https://github.com/suneeshjacob/ACRoD
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.0
