### 2023-10-10 suneeshjacob/ACRoD#2 @JHartzer
Update References
There are two references that likely need updating. I cannot find these papers myself, so if you could update these citations with links to the DOI number or preprint archive, that would be much appreciated!

- https://github.com/suneeshjacob/ACRoD/blob/f77170b9131359867dad473d774bea0dd607138e/examples/Jacobian/4R4P_parallel_robot.md?plain=1#L179
- https://github.com/suneeshjacob/ACRoD/blob/f77170b9131359867dad473d774bea0dd607138e/examples/Jacobian/RSSR-SSR_serial-parallel_hybrid_robot%20copy.md?plain=1#L183
