+++ older_version/src/jacobian_spatial.py
@@ -1,687 +1,0 @@
-import sympy
-import numpy
-import itertools
-
-def unique_list(input_list):
-    unique = []
-    for i in range(len(input_list)):
-        if input_list[i] not in unique:
-            unique.append(input_list[i])
-    return unique
-
-def jointslist(A):
-    n=len(A)
-    temp1=[]
-    temp2=[]
-    temp3=[]
-    temp4=[]
-    temp5=[]
-    for i in range(n):
-        for j in range(i+1,n):
-            if A[i,j]==1:
-                temp1.append((i,j,))
-            elif A[i,j]==2:
-                temp2.append((i,j,))
-            elif A[i,j]==3:
-                temp3.append((i,j,))
-            elif A[i,j]==4:
-                temp4.append((i,j,))
-            elif A[i,j]==5:
-                temp5.append((i,j,))
-#            temp.append(A[i,j])
-#    return [temp.count(1),temp.count(2),temp.count(3),temp.count(4)]
-    return [len(temp1),len(temp2),len(temp3),len(temp4),len(temp5)]
-
-def iscircular(list1,list2):
-    if len(list1)!=len(list2):
-        return False
-    temp1=' '.join(map(str,list1))
-    temp2=' '.join(map(str,list2))
-    if len(temp1)!=len(temp2):
-        return False
-    temp3=temp1+' '+temp1
-    if temp2 in temp3:
-        return True
-    else:
-        return False
-
-def reducedcyclic(valueslist):
-    for i in range(len(valueslist)):
-        if valueslist[i] in valueslist[i+1:]:
-            return valueslist[i:i+valueslist[i+1:].index(valueslist[i])+1]
-
-def listloops(A):
-    n=len(A)
-    a=[]
-    for i in range(n):
-        temp=[]
-        for j in list(range(0,i))+list(range(i+1,n)):
-            k=[j,i]
-            k.sort()
-            if A[k[0],k[1]]!=0:
-                temp.append(j)
-        a.append(temp)
-    
-#    temp=[]
-    temp5=[]
-    for i in range(n):
-        temp=[]
-        flag=0
-        temp.append([i])
-        if len(a[i])==0:
-            flag=1
-            ls=[]
-        else:
-            temp=[temp[0]+[i2] for i2 in a[i]]
-        if flag==0:
-            temp2=[]
-            for i3 in temp:
-                for i4 in a[i3[-1]]:
-                    if i4!=i3[-2]:
-                        temp2.append(i3+[i4])
-        flag2=0
-        while flag2==0:
-            temp2=[]
-            for i3 in temp:
-                for i4 in a[i3[-1]]:
-                    if i4!=i3[-2]:
-                        temp2.append(i3+[i4])
-            temp=temp2
-            flag2=1
-            for i4 in temp:
-                if len(set(i4))==len(i4):
-                    flag2=0
-                    break
-        if len(temp)!=0:
-            temp6=[]
-            for i5 in temp:
-                temp8=reducedcyclic(i5)
-                if len(temp8)==len(set(temp8)):
-                    temp6.append(temp8)
-            temp=temp6
-            temp5+=temp
-    if len(temp5)!=0:
-        temp7=[temp5[0]]
-        for i in range(len(temp5)):
-            for j in range(len(temp7)):
-                flag3=0
-                if iscircular(temp5[i],temp7[j])==True:
-                    flag3=1
-                    break
-                elif iscircular(temp5[i][::-1],temp7[j])==True:
-                    flag3=1
-                    break
-            if flag3==0:
-                temp7.append(temp5[i])
-    else:
-        temp7=temp5
-    return temp7
-
-def getalljoints(A,linknumber):
-    n=len(A)
[…]

[…]
