### 2023-10-10 suneeshjacob/ACRoD#3 @JHartzer
Increase Code Coverage
Excellent work on the unit tests. They are certainly a step in the right direction. My concern is they are simply checking for the original output and therefore don't share very much information. Additionally changes in the function output would be difficult to diagnose. Moving towards checking each array element one at a time would be more clear. Using a package like [unittest](https://docs.python.org/3/library/unittest.html) could also assist you as it has built-in functions for checking if output is near to an expected result.

Running [coverage.py](https://coverage.readthedocs.io/en/7.3.2/#) shows good coverage of `functions.py` at 86%. The main paths missing are all the `'descending'` options, `joint_type == 6`, and `joint_type == 7`. Getting those testing would put the package at nearly 100% code coverage.
