### 2023-10-10 suneeshjacob/ACRoD#4 @JHartzer
Remove Uses of numpy.matrix
From [numpy](https://numpy.org/doc/stable/reference/generated/numpy.matrix.html#numpy-matrix): "It is no longer recommended to use `numpy.matrix`, even for linear algebra. Instead use regular arrays. The class may be removed in the future."

With this depreciation warning in mind, it would be good to remove usage of the matrix class and replace the uses with `numpy.array`.
