### 2023-10-10 0111fa6
Remove .egg-info build artifacts
