+++ src/acrod.egg-info/SOURCES.txt
@@ -1,11 +1,0 @@
-LICENSE
-README.md
-pyproject.toml
-src/acrod/__init__.py
-src/acrod/functions.py
-src/acrod.egg-info/PKG-INFO
-src/acrod.egg-info/SOURCES.txt
-src/acrod.egg-info/dependency_links.txt
-src/acrod.egg-info/requires.txt
-src/acrod.egg-info/top_level.txt
-tests/test_jacobian.py
