### 2023-10-10 suneeshjacob/ACRoD#5 @JHartzer
Remove python package files
The package directories `dist` and `acrod.egg-info` should be removed as well as the old version of the code. Access previous versions of the code should be done through the git history.
