### 2023-10-22 e4fa11a
Update RSSR-SSR_serial-parallel_hybrid_robot copy.md
Removed a reference to an unpublished paper
