### 2023-10-22 5cdb86f
Update 4R4P_parallel_robot.md
Removed a reference to an unpublished paper
