### 2023-10-22 suneeshjacob/ACRoD#2 @suneeshjacob
Hi, the paper is an unpublished work by the same set of authors, and the cited items are only images. Due to some potential limitations of double blind peer review policies that wouldn't allow prior release of preprints, the authors decided to put this location as primary location for the images and the unpublished work, if published later, would cite this location appropriately. Changes relevant to this, are made in the repository. Hence, I believe this issue is resolved and so I am closing the issue.
