### 2023-10-22 67794a2
Update functions.py
Fixed a bug
