### 2023-10-22 c5ed203
Replaced matrix with array of NumPy
Replaced NumPy Matrix object with NumPy Array object and made appropriate changes
