### 2023-10-22 bcf41ac
Update functions.py
Fixed a bug
