### 2023-10-22 suneeshjacob/ACRoD#4 @suneeshjacob
Replaced matrix object of NumPy to array object in all occurrences of functions.py file, and the tests are successful. Hence the issue is resolved and i am closing the issue. Thanks.
