### 2023-10-24 1404d02
Added a test for helical joint
