+++ tests/test_jacobian.py
@@ -229,23 +229,25 @@ def test_spatial_R2SR():
 
 def test_spatial_3RPHS():
     A = numpy.array(
-        [[9, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0],
-         [0, 9, 6, 0, 0, 0, 0, 0, 0, 0, 0],
-         [0, 0, 9, 2, 0, 0, 0, 0, 0, 0, 0],
-         [0, 0, 1, 9, 0, 0, 0, 0, 0, 0, 4],
-         [0, 0, 0, 0, 9, 6, 0, 0, 0, 0, 0],
-         [0, 0, 0, 0, 0, 9, 2, 0, 0, 0, 0],
-         [0, 0, 0, 0, 0, 1, 9, 0, 0, 0, 4],
-         [0, 0, 0, 0, 0, 0, 0, 9, 6, 0, 0],
-         [0, 0, 0, 0, 0, 0, 0, 0, 9, 2, 0],
-         [0, 0, 0, 0, 0, 0, 0, 0, 1, 9, 4],
-         [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 9]]
+        [[9, 2, 0, 2, 0, 2, 0, 0],
+         [1, 9, 6, 0, 0, 0, 0, 0],
+         [0, 0, 9, 0, 0, 0, 0, 4],
+         [1, 0, 0, 9, 6, 0, 0, 0],
+         [0, 0, 0, 0, 9, 0, 0, 4],
+         [1, 0, 0, 0, 0, 9, 6, 0],
+         [0, 0, 0, 0, 0, 0, 9, 4],
+         [0, 0, 0, 0, 0, 0, 0, 9]]
        )
     jac = jacobian(A, robot_type = 'spatial')
     jac_func = jac.get_jacobian_function()
-    assert str([jac.Ja, jac.Jp, jac.Aa, jac.Ap]) == ''
[…]
