### 2023-10-24 6f28c9c
Added a test for plane joint and changed NumPy matrix objects to NumPy array objects
