+++ tests/test_jacobian.py
@@ -16,7 +16,7 @@ def test_planar_3Rs():
     end_effector_parameters = numpy.random.random(2)
     numpy.random.seed(0)
     robot_parameters = numpy.random.random(len(jac.parameters))
-    assert numpy.isclose(jac_func(end_effector_parameters, robot_parameters),numpy.matrix([[0.0, -0.17030618, -0.06929525], [0.0, -0.05394987, 0.1251587], [1.0, 1.0, 1.0]])).all()
+    assert numpy.isclose(jac_func(end_effector_parameters, robot_parameters),numpy.array([[0.0, -0.17030618, -0.06929525], [0.0, -0.05394987, 0.1251587], [1.0, 1.0, 1.0]])).all()
 
 def test_planar_3RPRp():
     A = numpy.array(
@@ -36,7 +36,7 @@ def test_planar_3RPRp():
     end_effector_parameters = numpy.random.random(2)
     numpy.random.seed(0)
     robot_parameters = numpy.random.random(len(jac.parameters))
-    assert numpy.isclose(jac_func(end_effector_parameters, robot_parameters),numpy.matrix([[1.5116862461994787, -1.001002195991253, 0.19757451735055986], [-0.41256134092451774, 1.381649776595281, -0.27270548341604917], [-9.331363937770648, 12.149247344303607, -1.3312812081939858]])).all()
+    assert numpy.isclose(jac_func(end_effector_parameters, robot_parameters),numpy.array([[1.5116862461994787, -1.001002195991253, 0.19757451735055986], [-0.41256134092451774, 1.381649776595281, -0.27270548341604917], [-9.331363937770648, 12.149247344303607, -1.3312812081939858]])).all()
 
     
 def test_planar_5Rp():
@@ -54,7 +54,7 @@ def test_planar_5Rp():
     end_effector_parameters = numpy.random.random(2)
     numpy.random.seed(0)
     robot_parameters = numpy.random.random(len(jac.parameters))
-    assert numpy.isclose(jac_func(end_effector_parameters, robot_parameters),numpy.matrix([[-0.07248152798410853, -0.059799857441136486], [-0.09063783901508533, 0.0842378131702819], [0.2184837927966302, 0.6668966302219953]])).all()
+    assert numpy.isclose(jac_func(end_effector_parameters, robot_parameters),numpy.array([[-0.07248152798410853, -0.059799857441136486], [-0.09063783901508533, 0.0842378131702819], [0.2184837927966302, 0.6668966302219953]])).all()
 
 
 def test_planar_4R4Psp():
@@ -74,7 +74,7 @@ def test_planar_4R4Psp():
     end_effector_parameters = numpy.random.random(2)
     numpy.random.seed(0)
     robot_parameters = numpy.random.random(len(jac.parameters))
-    assert numpy.isclose(jac_func(end_effector_parameters, robot_parameters),numpy.matrix([[0.7971752173945887, 2.4135074176540847, -2.878018570595505], [0.4831574969754604, 1.660978841059345, -3.3871377239062928], [0.0, 0.0, 9.466440213737489]])).all()
+    assert numpy.isclose(jac_func(end_effector_parameters, robot_parameters),numpy.array([[0.7971752173945887, 2.4135074176540847, -2.878018570595505], [0.4831574969754604, 1.660978841059345, -3.3871377239062928], [0.0, 0.0, 9.466440213737489]])).all()
 
 
 def test_spatial_6Rs():
@@ -94,7 +94,7 @@ def test_spatial_6Rs():
     end_effector_parameters = numpy.random.random(3)
     numpy.random.seed(0)
     robot_parameters = numpy.random.random(len(jac.parameters))
[…]
