### 2023-10-24 suneeshjacob/ACRoD#3 @suneeshjacob
Thanks for the suggestions. I added two more examples covering helical and plane joints (i.e., `joint_type == 6` and `joint_type == 7`). I will work on the remaining ones (i.e., `'descending'` and unittests) and will update about them.
