### 2023-10-24 suneeshjacob/ACRoD#7 @CameronDevine
Community Guidelines
As part of the JOSS review (openjournals/joss-reviews#5927) I have not found any community guidelines for,

1. contributing to the software,
2. reporting issues or problems, and
3. seeking support.
