### 2023-10-24 suneeshjacob/ACRoD#8 @CameronDevine
Statement of Need
The only documentation I have found so far is the [README.md](README.md) file. As part of the JOSS review (openjournals/joss-reviews#5927) this documentation should include a statement of need. I did not find anything I would consider a statement of need in the Readme.
