### 2023-10-24 0525c68
Update test_jacobian.py
Added a test to include planar `descending` in coverage.
