### 2023-10-24 #5927 @suneeshjacob
> Sorry for the delay, I've started checking the repo, I agree with the issues @JHartzer already opened there. I may be mistaken but it seems to me that, a part from some readme text, the repo is missing some example scripts or driving code. @suneeshjacob can you please check that you committed them or point me in the right direction? Thanks

Hi @gfadini, sorry i missed this comment. I have finished some of the issues that JHartzer opened, and I am working on the rest of them and I will notify once I finish all the issues. And regarding example scripts, yes the example is mentioned in the readme text. If needed, i can add more examples.
