### 2023-10-24 #5927 @suneeshjacob
> I am a little bit confused by the authorship of the paper. The only people who have contributed to the paper are @suneeshjacob and @JHartzer. It is not clear to me what work Rituparna Datta performed as part of this submission.

Hi @CameronDevine, many parts of the code, and especially the mathematical conceptualisation for completion of the code, have been worked upon under the guidance of Dr. Rituparna, and that's why his name is there as one of the authors in the paper.
