### 2023-10-25 #5927 @gfadini
> I am a little bit confused by the authorship of the paper. The only people who have contributed to the paper are @suneeshjacob and @JHartzer. It is not clear to me what work Rituparna Datta performed as part of this submission.

Hi @CameronDevine, please note that @JHartzer is one of the other reviewers. I think some of his suggestions were merged in the main repo.

Even if I do not think this is the case here, I wonder if this is generally fine for the JOSS review process itself, because it may create some conflict of interest (especially if the proposed modifications are numerous and significant).
Do you have any comments or thoughts on this point @Kevin-Mattheus-Moerman?
