### 2023-10-25 4a1e5c5
Update test_jacobian.py
Added a new function to include some instances of `descending` in coverage.
