+++ tests/test_jacobian.py
@@ -300,3 +300,41 @@ def test_planar_for_descending():
     assert numpy.isclose(jac_func(end_effector_parameters, robot_parameters),numpy.array([[-0.11882638,  0.50774852, -0.58812428],
        [-0.14723595,  0.34943317, -0.74914152],
        [ 0.        ,  0.        ,  2.37618934]])).all()
+
+def test_spatial_for_descending_1():
+    A = array(
+      [[9, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0],
+       [0, 9, 0, 0, 0, 0, 0, 0, 0, 0, 2, 0, 0, 0, 0, 4],
+       [0, 0, 9, 0, 0, 0, 4, 0, 0, 0, 0, 0, 0, 0, 2, 0],
+       [0, 0, 0, 9, 0, 0, 4, 2, 0, 0, 0, 0, 0, 0, 0, 0],
+       [0, 0, 0, 0, 9, 0, 0, 0, 0, 0, 0, 2, 0, 0, 0, 0],
+       [0, 0, 0, 0, 0, 9, 5, 0, 0, 0, 0, 0, 0, 0, 0, 2],
+       [0, 0, 0, 0, 0, 0, 9, 0, 0, 0, 4, 4, 4, 4, 0, 0],
+       [0, 0, 0, 1, 0, 0, 0, 9, 0, 0, 0, 0, 0, 0, 0, 4],
+       [0, 0, 0, 0, 0, 0, 0, 0, 9, 0, 0, 0, 0, 2, 0, 0],
+       [0, 0, 0, 0, 0, 0, 0, 0, 0, 9, 0, 0, 2, 0, 0, 4],
+       [0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 9, 0, 0, 0, 0, 0],
+       [0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 9, 0, 0, 0, 0],
+       [0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 9, 0, 0, 0],
+       [0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 9, 0, 0],
+       [0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 9, 0],
+       [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 9]])
+    jac = jacobian(A, robot_type = 'spatial')
+    jac_func = jac.get_jacobian_function()
[…]
