### 2023-10-25 dfaaf7f
Update test_jacobian.py
Added another test for including `descending` in coverage.
