+++ tests/test_jacobian.py
@@ -338,3 +338,34 @@ def test_spatial_for_descending_1():
          2.06073096e+00, -2.03709917e+00, -7.11180403e+01],
        [-5.74486260e+01, -2.71642977e+00,  8.60002598e+01,
          2.85012796e-01, -1.15294950e+00,  4.38284057e+02]])).all()
+
+
+def test_spatial_for_descending_2():
+    A = numpy.array(
+    [[9, 1, 0, 0, 0, 0, 0],
+     [1, 9, 1, 0, 0, 0, 0],
+     [0, 1, 9, 1, 0, 0, 0],
+     [0, 0, 1, 9, 1, 0, 0],
+     [0, 0, 0, 1, 9, 1, 0],
+     [0, 0, 0, 0, 1, 9, 1],
+     [0, 0, 0, 0, 0, 1, 9]]
+    )
+    jac = jacobian(A, robot_type = 'spatial')
+    jac_func = jac.get_jacobian_function()
+    assert str([jac.Ja, jac.Jp, jac.Aa, jac.Ap]) == '[Matrix([\n[                  -(a_y - r_{(1,6)y})*cos(\\beta_{(1,6)}) + (a_z - r_{(1,6)z})*sin(\\beta_{(1,6)})*sin(\\phi_{(1,6)}),                   -(a_y - r_{(2,4)y})*cos(\\beta_{(2,4)}) + (a_z - r_{(2,4)z})*sin(\\beta_{(2,4)})*sin(\\phi_{(2,4)}),                   (a_y - r_{(2,6)y})*cos(\\beta_{(2,6)}) - (a_z - r_{(2,6)z})*sin(\\beta_{(2,6)})*sin(\\phi_{(2,6)}),                   (a_y - r_{(3,4)y})*cos(\\beta_{(3,4)}) - (a_z - r_{(3,4)z})*sin(\\beta_{(3,4)})*sin(\\phi_{(3,4)}),                   -(a_y - r_{(3,5)y})*cos(\\beta_{(3,5)}) + (a_z - r_{(3,5)z})*sin(\\beta_{(3,5)})*sin(\\phi_{(3,5)}),                   -(a_y - r_{(5,7)y})*cos(\\beta_{(5,7)}) + (a_z - r_{(5,7)z})*sin(\\beta_{(5,7)})*sin(\\phi_{(5,7)})],\n[                   (a_x - r_{(1,6)x})*cos(\\beta_{(1,6)}) - (a_z - r_{(1,6)z})*sin(\\beta_{(1,6)})*cos(\\phi_{(1,6)}),                    (a_x - r_{(2,4)x})*cos(\\beta_{(2,4)}) - (a_z - r_{(2,4)z})*sin(\\beta_{(2,4)})*cos(\\phi_{(2,4)}),                  -(a_x - r_{(2,6)x})*cos(\\beta_{(2,6)}) + (a_z - r_{(2,6)z})*sin(\\beta_{(2,6)})*cos(\\phi_{(2,6)}),                  -(a_x - r_{(3,4)x})*cos(\\beta_{(3,4)}) + (a_z - r_{(3,4)z})*sin(\\beta_{(3,4)})*cos(\\phi_{(3,4)}),                    (a_x - r_{(3,5)x})*cos(\\beta_{(3,5)}) - (a_z - r_{(3,5)z})*sin(\\beta_{(3,5)})*cos(\\phi_{(3,5)}),                    (a_x - r_{(5,7)x})*cos(\\beta_{(5,
[…]
