### 2023-10-26 4f925c1
Update test_jacobian.py
Added another test to include `descending` for cylindrical joint.
