+++ tests/test_jacobian.py
@@ -368,3 +368,27 @@ def test_spatial_for_descending_2():
          0.20082922],
        [ 0.85518784,  0.57051564, -0.601358  , -0.71220865,  0.71054128,
          0.8668993 ]])).all()
+
+def test_spatial_for_descending_3():
+    A = numpy.array(
+      [[9, 0, 0, 0, 2, 2, 2, 0],
+       [0, 9, 0, 0, 0, 0, 3, 5],
+       [0, 0, 9, 0, 0, 3, 0, 5],
+       [0, 0, 0, 9, 3, 0, 0, 5],
+       [1, 0, 0, 0, 9, 0, 0, 0],
+       [1, 0, 0, 0, 0, 9, 0, 0],
+       [1, 0, 0, 0, 0, 0, 9, 0],
+       [0, 0, 0, 0, 0, 0, 0, 9]])
+    jac = jacobian(A, robot_type = 'spatial')
+    jac_func = jac.get_jacobian_function()
+    assert str([jac.Ja, jac.Jp, jac.Aa, jac.Ap]) == '[Matrix([\n[sin(\\beta_{(1,5)})*cos(\\phi_{(1,5)}), 0, 0],\n[sin(\\beta_{(1,5)})*sin(\\phi_{(1,5)}), 0, 0],\n[                  cos(\\beta_{(1,5)}), 0, 0],\n[                                   0, 0, 0],\n[                                   0, 0, 0],\n[                                   0, 0, 0]]), Matrix([\n[0, 0, 0, 0, 0, 0, 0, 0,                   (a_y - r_{(4,5)y})*cos(\\beta_{(4,5)}) - (a_z - r_{(4,5)z})*sin(\\beta_{(4,5)})*sin(\\phi_{(4,5)}), -sin(\\beta_{(4,5)})*cos(\\phi_{(4,5)}),                   -(a_y - r_{(4,8)y})*cos(\\beta_{(4,8)}) + (a_z - r_{(4,8)z})*sin(\\beta_{(4,8)})*sin(\\phi_{(4,8)}),                                                                            -(a_y - r_{(4,8)y})/sqrt(1 + 1/(cos(\\delta_{(4,8)} - \\phi_{(4,8)})**2*tan(\\beta_{(4,8)})**2)) - (a_z - r_{(4,8)z})*sin(\\delta_{(4,8)})/(sqrt(1 + 1/(cos(\\delta_{(4,8)} - \\phi_{(4,8)})**2*tan(\\beta_{(4,8)})**2))*cos(\\delta_{(4,8)} - \\phi_{(4,8)})*tan(\\beta_{(4,8)}))],\n[0, 0, 0, 0, 0, 0, 0, 0,                  -(a_x - r_{(4,5)x})*cos(\\beta_{(4,5)}) + (a_z - r_{(4,5)z})*sin(\\beta_{(4,5)})*cos(\\phi_{(4,5)}), -sin(\\beta_{(4,5)})*sin(\\phi_{(4,5)}),                    (a_x - r_{(4,8)x})*cos(\\beta_{(4,8)}) - (a_z - r_{(4,8)z})*sin(\\beta_{(4,8)})*cos(\\phi_{(4,8)}),                                                                             (a_x - r_{(4,8)x})/sqrt(1 + 1/(cos(\\delta_{(4,8)} - \\phi_{(4,8)})**2*tan(\\beta_{(4,8)})**2)) + (a_z - r_{(4,8)z})*cos(\
[…]
