### 2023-10-26 58eff68
Update test_jacobian.py
Added another example to include `descending` for helical joint in coverage
