+++ tests/test_jacobian.py
@@ -418,3 +418,27 @@ def test_spatial_for_descending_4():
        [-1.30443781, -0.33612152,  1.31540466],
        [-0.74871083,  1.02885713,  0.06152439],
        [ 1.14290989, -1.34331174, -0.40804821]])).all()
+
+def test_spatial_for_descending_5():
+    A = numpy.array(
+      [[9, 1, 0, 0, 0, 1, 1, 0],
+       [1, 9, 0, 0, 7, 0, 0, 0],
+       [0, 0, 9, 0, 0, 7, 0, 4],
+       [0, 0, 0, 9, 0, 0, 7, 4],
+       [0, 0, 0, 0, 9, 0, 0, 4],
+       [1, 0, 0, 0, 0, 9, 0, 0],
+       [1, 0, 0, 0, 0, 0, 9, 0],
+       [0, 0, 0, 0, 0, 0, 0, 9]])
+    jac = jacobian(A, robot_type = 'spatial')
+    jac_func = jac.get_jacobian_function()
+    assert str([jac.Ja, jac.Jp, jac.Aa, jac.Ap]) == '[Matrix([\n[                  -(a_y - r_{(1,2)y})*cos(\\beta_{(1,2)}) + (a_z - r_{(1,2)z})*sin(\\beta_{(1,2)})*sin(\\phi_{(1,2)}), 0, 0],\n[                   (a_x - r_{(1,2)x})*cos(\\beta_{(1,2)}) - (a_z - r_{(1,2)z})*sin(\\beta_{(1,2)})*cos(\\phi_{(1,2)}), 0, 0],\n[-(a_x - r_{(1,2)x})*sin(\\beta_{(1,2)})*sin(\\phi_{(1,2)}) + (a_y - r_{(1,2)y})*sin(\\beta_{(1,2)})*cos(\\phi_{(1,2)}), 0, 0],\n[                                                                              sin(\\beta_{(1,2)})*cos(\\phi_{(1,2)}), 0, 0],\n[                                                                              sin(\\beta_{(1,2)})*sin(\\phi_{(1,2)}), 0, 0],\n[                                                                                                cos(\\beta_{(1,2)}), 0, 0]]), Matrix([\n[sin(\\beta_{(2,5)})*cos(\\phi_{(2,5)}), 0, 0, 0, 0, 0, 0, 0, 0,                 0,  a_z - r_{(5,8)z}, -a_y + r_{(5,8)y}],\n[sin(\\beta_{(2,5)})*sin(\\phi_{(2,5)}), 0, 0, 0, 0, 0, 0, 0, 0, -a_z + r_{(5,8)z},                 0,  a_x - r_{(5,8)x}],\n[                  cos(\\beta_{(2,5)}), 0, 0, 0, 0, 0, 0, 0, 0,  a_y - r_{(5,8)y}, -a_x + r_{(5,8)x},                 0],\n[                                   0, 0, 0, 0, 0, 0, 0, 0, 0,                 1,                 0,                 0],\n[                                   0, 0, 0, 0, 0, 0, 0, 0, 0,                 0,                 1,                 0],\n[                                   0, 0, 0, 0, 0, 0, 0, 0, 0,                 0,                 0,                 1]]), Matrix([\n[                  (a_y - r_{(1,2)y})*cos(\\beta_{(1,2)}) - (a_z - r_{(1,2)z})*sin(\\beta_{(1,2)})*sin(\\phi_{(1,2)}),                   -(a_y - r_{(1,6)y})*cos(\\beta_{(1,6)}) + (a_z - r_{(1,6)z})*sin(\\beta_{(1,6)})*sin(\\phi_{(1,6)}),                                                                                                                  0],\n[                 -(a_x - r_{(1,2)x})*cos(\\beta_{(1,2)}) + (a_z - r_{
[…]
