+++ tests/test_jacobian.py
@@ -396,28 +396,27 @@ def test_spatial_for_descending_3():
 
 def test_spatial_for_descending_4():
     A = numpy.array(
-        [[9, 2, 0, 2, 0, 2, 0, 0],
-         [1, 9, 6, 0, 0, 0, 0, 0],
-         [0, 0, 9, 0, 0, 0, 0, 4],
-         [1, 0, 0, 9, 6, 0, 0, 0],
-         [0, 0, 0, 0, 9, 0, 0, 4],
-         [1, 0, 0, 0, 0, 9, 6, 0],
-         [0, 0, 0, 0, 0, 0, 9, 4],
-         [0, 0, 0, 0, 0, 0, 0, 9]]
-       )
+      [[9, 0, 2, 0, 2, 0, 2, 0],
+       [0, 9, 0, 0, 0, 0, 6, 4],
+       [1, 0, 9, 0, 0, 6, 0, 0],
+       [0, 0, 0, 9, 6, 0, 0, 4],
+       [1, 0, 0, 0, 9, 0, 0, 0],
+       [0, 0, 0, 0, 0, 9, 0, 4],
+       [1, 0, 0, 0, 0, 0, 9, 0],
+       [0, 0, 0, 0, 0, 0, 0, 9]])
     jac = jacobian(A, robot_type = 'spatial')
     jac_func = jac.get_jacobian_function()
-    assert str([jac.Ja, jac.Jp, jac.Aa, jac.Ap]) == '[Matrix([\n[sin(\\beta_{(1,5)})*cos(\\phi_{(1,5)}), 0, 0],\n[sin(\\beta_{(1,5)})*sin(\\phi_{(1,5)}), 0, 0],\n[                  cos(\\beta_{(1,5)}), 0, 0],\n[                                   0, 0, 0],\n[                                   0, 0, 0],\n[                                   0, 0, 0]]), Matrix([\n[0, 0, 0, 0, -p_{(3,5)}*sin(\\beta_{(3,5)})*cos(\\phi_{(3,5)})/(2*pi) + (a_y - r_{(3,5)y})*cos(\\beta_{(3,5)}) - (a_z - r_{(3,5)z})*sin(\\beta_{(3,5)})*sin(\\phi_{(3,5)}),                 0,  a_z - r_{(3,8)z}, -a_y + r_{(3,8)y}, 0, 0, 0, 0],\n[0, 0, 0, 0, -p_{(3,5)}*sin(\\beta_{(3,5)})*sin(\\phi_{(3,5)})/(2*pi) - (a_x - r_{(3,5)x})*cos(\\beta_{(3,5)}) + (a_z - r_{(3,5)z})*sin(\\beta_{(3,5)})*cos(\\phi_{(3,5)}), -a_z + r_{(3,8)z},                 0,  a_x - r_{(3,8)x}, 0, 0, 0, 0],\n[0, 0, 0, 0, -p_{(3,5)}*cos(\\beta_{(3,5)})/(2*pi) + (a_x - r_{(3,5)x})*sin(\\beta_{(3,5)})*sin(\\phi_{(3,5)}) - (a_y - r_{(3,5)y})*sin(\\beta_{(3,5)})*cos(\\phi_{(3,5)}),  a_y - r_{(3,8)y}, -a_x + r_{(3,8)x},                 0, 0, 0, 0, 0],\n[0, 0, 0, 0,                                                                                                                    -sin(\\beta_{(3,5)})*cos(\\phi_{(3,5)}),                 1,                 0,                 0, 0, 0, 0, 0],\n[0, 0, 0, 0,                                                                                                                    -sin(\\beta_{(3,5)})*sin(\\phi_{(3,5)}),                 0,                 1,                 0, 0, 0, 0, 0],\n[0, 0, 0, 0,                                                                                   
[…]
