### 2023-10-26 suneeshjacob/ACRoD#3 @suneeshjacob
I added a few functions covering `descending` and now coverage.py is showing the coverage of `functions.py` at 98%.
