### 2023-10-26 suneeshjacob/ACRoD#3 @suneeshjacob
@JHartzer Regarding checking output, each test function checks both the output string and also the jacobian matrix for a random set of dimensional parameters of the robot at a random end-effector point (by using `numpy.isclose`. The output string comparison is to ensure that the symbolic matrix functions for Jacobian are being generated properly, and the other comparison is to ensure numerical values of jacobian matrix for a given pose. So, both numerical and symbolic results are tested in the tests and they include comparison of the entire matrices, which include individual comparisons as well, as they would pass the tests only if individual elements are close enough. Is a separate comparison of individual elements of matrices still required?
