### 2023-10-26 #5927 @suneeshjacob
> > > Sorry for the delay, I've started checking the repo, and I agree with the issues @JHartzer already opened there. I may be mistaken but it seems to me that, apart from some readme text, the repo is missing some example scripts or driving code. @suneeshjacob can you please check that you committed them or point me in the right direction? Thanks
> >
> >
> > Hi @gfadini, sorry i missed this comment. I have finished some of the issues that JHartzer opened, and I am working on the rest of them and I will notify once I finish all the issues. And regarding example scripts, yes the example is mentioned in the readme text. If needed, i can add more examples.
>
> Thanks for the reply and for confirming my point. I think it would be great if you could add more examples of use cases. The repo `readme` shows how to use the library in a rather simple case. Then you worked out the math behind some jacobian formulations in `[examples/Jacobian](https://github.com/suneeshjacob/ACRoD/tree/main/examples/Jacobian)`. However, you do not show how to use `ACRoD` to obtain those expressions. I find this rather strange, as I would expect to have `examples` of use rather than just the mathematical part (which should just be there to double-check the output). Can you please work on this?

Yeah i'll work on it and will give you an update on this. Thanks.
