### 2023-10-27 e5a0143
Update 3R_serial_robot.md
Updated figure format
