### 2023-10-27 0553735
Update 3R_serial_robot.md
Fixed a bug
