### 2023-10-27 9d29ce5
Update 3R_serial_robot.md
Reverted back to the previous image format
