### 2023-10-27 suneeshjacob/ACRoD#8 @suneeshjacob
Yes, the documentation is currently only in [README.md](https://github.com/suneeshjacob/ACRoD/issues/README.md) file (and the full API documentation is aimed for future version). And I am already planning to add more examples based on `[gfadini](https://github.com/gfadini)`'s review. A statement of need was included in the paper but was not included in the documentation. It is now included in the [README.md](https://github.com/suneeshjacob/ACRoD/issues/README.md) file, thanks for pointing that out.
