### 2023-10-28 suneeshjacob/ACRoD#7 @suneeshjacob
I now added community guidelines at the bottom of [README.md](https://github.com/suneeshjacob/ACRoD/blob/main/README.md).
