+++ README.md
@@ -95,6 +95,9 @@ $$\left[\begin{matrix}a_{x} \\\\ a_{y}\end{matrix}\right]$$
 
 #### Sample computation of Jacobian at the end-effector point $\textbf{a}=\hat{i}+2\hat{j}$ and at the configuration of $\textbf{r}\_{(1,2)}=3\hat{i}+4\hat{j}$ and $\textbf{r}\_{(2,3)}=5\hat{i}+6\hat{j}$
 
+- End-effector point: $\textbf{a}=\hat{i}+2\hat{j}$
+- Locations of joints: $\textbf{r}\_{(1,2)}=3\hat{i}+4\hat{j}$ and $\textbf{r}\_{(2,3)}=5\hat{i}+6\hat{j}$
+
 For the given set of dimensional parameters of the robot, the numerical Jacobian can be computed as follows.
 ```py
 end_effector_point = [1,2]
