### 2023-10-28 d2fe404
Added a file template for 4R4P example
