+++ 4R4P.md
@@ -1,0 +1,107 @@
+# Formulation of Jacobian for 4R-4P serial-parallel hybrid manipulator as an example
+
+## Usage
+
+### Jacobian for planar manipulators
+
+The topological information of a robot is to be specified by using its robot-topology matrix, as defined [here](Robot_Topology_Matrix.md). For a planar 2R serial manipulator, the robot topology matrix is given by
+
+$$\left[\begin{matrix}
+9 & 1 & 0 \\\\
+1 & 9 & 1 \\\\
+0 & 1 & 9
+\end{matrix}\right]$$
+
+The corresponding Jacobian function can be formulated as follows.
+
+Firstly, the required functions are imported as shown below.
+```py
+from acrod.functions import jacobian
+from numpy import matrix
+```
+
+
+The robot-topology matrix for 3R planar serial manipulator is defined and jacobian information is processed via the imported jacobian class as follows.
+```py
+M = matrix('9 1 0;1 9 1;0 1 9')
+jac = jacobian(M, robot_type = 'planar')
+```
+
+
+Jacobian function is generated as shown below.
+```py
+jacobian_function = jac.get_jacobian_function()
+```
+
+
+In the process of generating the above jacobian function, other attributes of the jacobian object also are updated. Symbolic Jacobian matrices can be extracted from the attributes. Since this is a serial robot, the matrix $J_a$ itself would be the Jacobian matrix of the manipulator. The matrix $J_a$ is extracted from `Ja` attribute of the jacobian object as follows.
+```py
+symbolic_jacobian = jac.Ja
+symbolic_jacobian
+```
+
+In an ipynb file of JupyterLab, the above code would produce the following output.
+
+$$\left[\begin{matrix}- a\_{y} + r\_{(1,2)y} & - a\_{y} + r\_{(2,3)y} \\\\ a\_{x} - r\_{(1,2)x} & a\_{x} - r\_{(2,3)x} \\\\ 1 & 1\end{matrix}\right]$$
+
+The above Jacobian is based on the notations defined and described [here](Notation_and_Nomenclature.md).
+
+Active joint velocities, in the corresponding order, can be viewed by running the following lines.
+```py
+active_joint_velocities = jac.active_joint_velocities_symbolic
+active_joint_velocities
+```
+
+In an ipynb file of JupyterLab, the above code would produce the following output.
+
+$$\left[\begin{matrix}\dot{\theta}\_{(1,2)} \\\\ \dot{\theta}\_{(2,3)}\end{matrix}\right]$$
+
+Robot dimensional parameters can be viewed by running the below line.
+```py
+robot_dimensional_parameters = jac.parameters_symbolic
+robot_dimensional_parameters
+```
+
+In an ipynb file of JupyterLab, the above code would produce the following output.
+
+$$\left[\begin{matrix}r_{(1,2)x} \\\\ r_{(1,2)y} \\\\ r_{(2,3)x} \\\\ r_{(2,3)y}\end{matrix}\right]$$
+
+
+Robot end-effector parameters can be viewed by running the below line.
+```py
+robot_endeffector_parameters = jac.endeffector_variables_symbolic
+robot_endeffector_parameters
+```
+
+In an ipynb file of JupyterLab, the above code would produce the following output.
+
+$$\left[\begin{matrix}a_{x} \\\\ a_{y}\end{matrix}\right]$$
+
+#### Sample computation of Jacobian for the configuration corresponding to the parameters shown below:
+
+- End-effector point: $\textbf{a}=\hat{i}+2\hat{j}$
+- Locations of joints: $\textbf{r}\_{(1,2)}=3\hat{i}+4\hat{j}$ and $\textbf{r}\_{(2,3)}=5\hat{i}+6\hat{j}$
+
[…]
