+++ 4R4P.md
@@ -1,5 +1,15 @@
 # Formulation of Jacobian for 4R-4P serial-parallel hybrid manipulator as an example
 
+## Figure
+
+<p align="center">
+    <img src="./RRRRPPPP.png" alt="4R-4P planar serial-parallel hybrid manipulator" width="500px">
+</p>
+
+A figure of RRRRPPPP planar serial-parallel hybrid manipulator is shown above. The corresponding adjacency matrix is given by
+
+$$\bf{M} = \left[\begin{matrix}L_1 & P & P & O & O & O & O\\\\A & L_2 & O & P & O & O & O\\\\A & O & L_3 & P & O & O & O\\\\O & O & O & L_4 & R & R & O\\\\O & O & O & A & L_5 & O & R\\\\O & O & O & O & O & L_6 & R\\\\O & O & O & O & O & O & L_7\end{matrix}\right]$$
+
 ## Usage
 
 ### Jacobian for planar manipulators
