+++ 4R4P.md
@@ -23,13 +23,21 @@ The corresponding Jacobian function can be formulated as follows.
 Firstly, the required functions are imported as shown below.
 ```py
 from acrod.functions import jacobian
-from numpy import matrix
+from numpy import array
 ```
 
 
 The robot-topology matrix for 3R planar serial manipulator is defined and jacobian information is processed via the imported jacobian class as follows.
 ```py
-M = matrix('9 1 0;1 9 1;0 1 9')
+M = numpy.array(
+        [[9, 2, 2, 0, 0, 0, 0],
+         [1, 9, 0, 2, 0, 0, 0],
+         [1, 0, 9, 2, 0, 0, 0],
+         [0, 0, 0, 9, 1, 1, 0],
+         [0, 0, 0, 1, 9, 0, 1],
+         [0, 0, 0, 0, 0, 9, 1],
+         [0, 0, 0, 0, 0, 0, 9]]
+    )
 jac = jacobian(M, robot_type = 'planar')
 ```
 
