+++ 4R4P.md
@@ -48,7 +48,28 @@ jacobian_function = jac.get_jacobian_function()
 ```
 
 
-In the process of generating the above jacobian function, other attributes of the jacobian object also are updated. Symbolic Jacobian matrices can be extracted from the attributes. Since this is a serial robot, the matrix $J_a$ itself would be the Jacobian matrix of the manipulator. The matrix $J_a$ is extracted from `Ja` attribute of the jacobian object as follows.
+In the process of generating the above jacobian function, other attributes of the jacobian object also are updated. Symbolic Jacobian matrices can be extracted from the attributes. Since this is a non-serial robot, there would be four matrices required to compute the Jacobian, which are $J_a$, $J_p$, $A_a$ and $A_p$. These can be extracted by the attributes `Ja`, `Jp`, `Aa` and `Ap`, respectively, as shown below.
+
+```py
+symbolic_Ja = jac.Ja
+symbolic_Ja
+```
+
+```py
+symbolic_Jp = jac.Jp
+symbolic_Jp
+```
+
+```py
+symbolic_Aa = jac.Aa
+symbolic_Aa
+```
+```py
+symbolic_Ap = jac.Ap
+symbolic_Ap
+```
+
+Since this is a serial robot, the matrix $J_a$ itself would be the Jacobian matrix of the manipulator. The matrix $J_a$ is extracted from `Ja` attribute of the jacobian object as follows.
 ```py
 symbolic_jacobian = jac.Ja
 symbolic_jacobian
