+++ 4R4P.md
@@ -56,6 +56,7 @@ symbolic_Ja
 ```
 
 Output in Jupyter notebook:
+
 $$\left[\begin{matrix}\cos{\left(\phi_{(1,2)} \right)} & 0 & - a_{y} + r_{(4,5)y}\\\sin{\left(\phi_{(1,2)} \right)} & 0 & a_{x} - r_{(4,5)x}\\0 & 0 & 1\end{matrix}\right]$$
 
 ```py
