+++ 4R4P.md
@@ -63,15 +63,24 @@ $$\left[\begin{matrix}\cos{\left(\phi_{(1,2)} \right)} & 0 & - a_{y} + r_{(4,5)y
 symbolic_Jp = jac.Jp
 symbolic_Jp
 ```
+Output in Jupyter notebook:
+
+$$\left[\begin{matrix}\cos{\left(\phi_{(1,2)} \right)} & 0 & - a_{y} + r_{(4,5)y}\\\\\sin{\left(\phi_{(1,2)} \right)} & 0 & a_{x} - r_{(4,5)x}\\\\0 & 0 & 1\end{matrix}\right]$$
 
 ```py
 symbolic_Aa = jac.Aa
 symbolic_Aa
 ```
+Output in Jupyter notebook:
+
+$$\left[\begin{matrix}\cos{\left(\phi_{(1,2)} \right)} & 0 & - a_{y} + r_{(4,5)y}\\\\\sin{\left(\phi_{(1,2)} \right)} & 0 & a_{x} - r_{(4,5)x}\\\\0 & 0 & 1\end{matrix}\right]$$
 ```py
 symbolic_Ap = jac.Ap
 symbolic_Ap
 ```
+Output in Jupyter notebook:
+
+$$\left[\begin{matrix}\cos{\left(\phi_{(1,2)} \right)} & 0 & - a_{y} + r_{(4,5)y}\\\\\sin{\left(\phi_{(1,2)} \right)} & 0 & a_{x} - r_{(4,5)x}\\\\0 & 0 & 1\end{matrix}\right]$$
 
 Since this is a serial robot, the matrix $J_a$ itself would be the Jacobian matrix of the manipulator. The matrix $J_a$ is extracted from `Ja` attribute of the jacobian object as follows.
 ```py
