+++ 4R4P.md
@@ -65,7 +65,7 @@ symbolic_Jp
 ```
 Output in Jupyter notebook:
 
-$$\left[\begin{matrix}\cos{\left(\phi_{(1,2)} \right)} & 0 & - a_{y} + r_{(4,5)y}\\\\\sin{\left(\phi_{(1,2)} \right)} & 0 & a_{x} - r_{(4,5)x}\\\\0 & 0 & 1\end{matrix}\right]$$
+$$\left[\begin{matrix}\cos{\left(\phi_{(2,4)} \right)} & 0 & 0 & - a_{y} + r_{(5,7)y} & 0\\\sin{\left(\phi_{(2,4)} \right)} & 0 & 0 & a_{x} - r_{(5,7)x} & 0\\0 & 0 & 0 & 1 & 0\end{matrix}\right]$$
 
 ```py
 symbolic_Aa = jac.Aa
@@ -73,14 +73,14 @@ symbolic_Aa
 ```
 Output in Jupyter notebook:
 
-$$\left[\begin{matrix}\cos{\left(\phi_{(1,2)} \right)} & 0 & - a_{y} + r_{(4,5)y}\\\\\sin{\left(\phi_{(1,2)} \right)} & 0 & a_{x} - r_{(4,5)x}\\\\0 & 0 & 1\end{matrix}\right]$$
+$$\left[\begin{matrix}0 & 0 & a_{y} - r_{(4,5)y}\\0 & 0 & - a_{x} + r_{(4,5)x}\\- \cos{\left(\phi_{(1,2)} \right)} & \cos{\left(\phi_{(1,3)} \right)} & 0\\- \sin{\left(\phi_{(1,2)} \right)} & \sin{\left(\phi_{(1,3)} \right)} & 0\\0 & 0 & -1\end{matrix}\right]$$
 ```py
 symbolic_Ap = jac.Ap
 symbolic_Ap
 ```
 Output in Jupyter notebook:
 
-$$\left[\begin{matrix}\cos{\left(\phi_{(1,2)} \right)} & 0 & - a_{y} + r_{(4,5)y}\\\\\sin{\left(\phi_{(1,2)} \right)} & 0 & a_{x} - r_{(4,5)x}\\\\0 & 0 & 1\end{matrix}\right]$$
+$$\left[\begin{matrix}0 & 0 & - a_{y} + r_{(4,6)y} & a_{y} - r_{(5,7)y} & - a_{y} + r_{(6,7)y}\\0 & 0 & a_{x} - r_{(4,6)x} & - a_{x} + r_{(5,7)x} & a_{x} - r_{(6,7)x}\\- \cos{\left(\phi_{(2,4)} \right)} & \cos{\left(\phi_{(3,4)} \right)} & 0 & 0 & 0\\- \sin{\left(\phi_{(2,4)} \right)} & \sin{\left(\phi_{(3,4)} \right)} & 0 & 0 & 0\\0 & 0 & 1 & -1 & 1\end{matrix}\right]$$
 
 Since this is a serial robot, the matrix $J_a$ itself would be the Jacobian matrix of the manipulator. The matrix $J_a$ is extracted from `Ja` attribute of the jacobian object as follows.
 ```py
