+++ 4R4P.md
@@ -73,7 +73,7 @@ symbolic_Aa
 ```
 Output in Jupyter notebook:
 
-$$\left[\begin{matrix}0 & 0 & a_{y} - r_{(4,5)y}\\\\0 & 0 & - a_{x} + r_{(4,5)x}\\- \cos{\left(\phi_{(1,2)} \right)} & \cos{\left(\phi_{(1,3)} \right)} & 0\\\\- \sin{\left(\phi_{(1,2)} \right)} & \sin{\left(\phi_{(1,3)} \right)} & 0\\\\0 & 0 & -1\end{matrix}\right]$$
+$$\left[\begin{matrix}0 & 0 & a_{y} - r_{(4,5)y}\\\\0 & 0 & - a_{x} + r_{(4,5)x}\\\\- \cos{\left(\phi_{(1,2)} \right)} & \cos{\left(\phi_{(1,3)} \right)} & 0\\\\- \sin{\left(\phi_{(1,2)} \right)} & \sin{\left(\phi_{(1,3)} \right)} & 0\\\\0 & 0 & -1\end{matrix}\right]$$
 ```py
 symbolic_Ap = jac.Ap
 symbolic_Ap
