+++ 4R4P.md
@@ -82,7 +82,7 @@ Output in Jupyter notebook:
 
 $$\left[\begin{matrix}0 & 0 & - a_{y} + r_{(4,6)y} & a_{y} - r_{(5,7)y} & - a_{y} + r_{(6,7)y}\\\\0 & 0 & a_{x} - r_{(4,6)x} & - a_{x} + r_{(5,7)x} & a_{x} - r_{(6,7)x}\\\\- \cos{\left(\phi_{(2,4)} \right)} & \cos{\left(\phi_{(3,4)} \right)} & 0 & 0 & 0\\\\- \sin{\left(\phi_{(2,4)} \right)} & \sin{\left(\phi_{(3,4)} \right)} & 0 & 0 & 0\\\\0 & 0 & 1 & -1 & 1\end{matrix}\right]$$
 
-$$\left[\begin{matrix}0 & 0 & \frac{\sin{\left(\phi_{(3,4)} \right)}}{\sin{\left(\phi_{(2,4)} - \phi_{(3,4)} \right)}} & - \frac{\cos{\left(\phi_{(3,4)} \right)}}{\sin{\left(\phi_{(2,4)} - \phi_{(3,4)} \right)}} & 0\\\\0 & 0 & \frac{\sin{\left(\phi_{(2,4)} \right)}}{\sin{\left(\phi_{(2,4)} - \phi_{(3,4)} \right)}} & - \frac{\cos{\left(\phi_{(2,4)} \right)}}{\sin{\left(\phi_{(2,4)} - \phi_{(3,4)} \right)}} & 0\\\\\frac{- r_{(5,7)x} + r_{(6,7)x}}{r_{(4,6)x} r_{(5,7)y} - r_{(4,6)x} r_{(6,7)y} - r_{(4,6)y} r_{(5,7)x} + r_{(4,6)y} r_{(6,7)x} + r_{(5,7)x} r_{(6,7)y} - r_{(5,7)y} r_{(6,7)x}} & \frac{- r_{(5,7)y} + r_{(6,7)y}}{r_{(4,6)x} r_{(5,7)y} - r_{(4,6)x} r_{(6,7)y} - r_{(4,6)y} r_{(5,7)x} + r_{(4,6)y} r_{(6,7)x} + r_{(5,7)x} r_{(6,7)y} - r_{(5,7)y} r_{(6,7)x}} & 0 & 0 & \frac{a_{x} r_{(5,7)y} - a_{x} r_{(6,7)y} - a_{y} r_{(5,7)x} + a_{y} r_{(6,7)x} + r_{(5,7)x} r_{(6,7)y} - r_{(5,7)y} r_{(6,7)x}}{r_{(4,6)x} r_{(5,7)y} - r_{(4,6)x} r_{(6,7)y} - r_{(4,6)y} r_{(5,7)x} + r_{(4,6)y} r_{(6,7)x} + r_{(5,7)x} r_{(6,7)y} - r_{(5,7)y} r_{(6,7)x}}\\\\\frac{- r_{(4,6)x} + r_{(6,7)x}}{r_{(4,6)x} r_{(5,7)y} - r_{(4,6)x} r_{(6,7)y} - r_{(4,6)y} r_{(5,7)x} + r_{(4,6)y} r_{(6,7)x} + r_{(5,7)x} r_{(6,7)y} - r_{(5,7)y} r_{(6,7)x}} & \frac{- r_{(4,6)y} + r_{(6,7)y}}{r_{(4,6)x} r_{(5,7)y} - r_{(4,6)x} r_{(6,7)y} - r_{(4,6)y} r_{(5,7)x} + r_
[…]
