### 2023-10-31 suneeshjacob/ACRoD#3 @JHartzer
I'm glad to hear you were able to increase the code coverage and test the additional joint types. Regarding your question, I think it is a good practice to check in such a way that if you change your code and a test fails, it is clear what has failed. However, if you don't foresee the output changing drastically, then I don't think you should go through the effort of refactoring these tests and can consider this issue closed with the increased code coverage.
