### 2023-10-31 suneeshjacob/ACRoD#9 @JHartzer
JOSS: Paper Comments
This issue will be a collection of comments regarding the JOSS paper itself.

# Summary
Overall, I believe the summary is good, but think the comparisons to other software should be extended and moved to the statement of need.

# Statement of Need
Extend the comparison to the other software packages and their limitations to make the statement of need stronger.

Additionally, I wonder if the second paragraph deserves its own section, like`Method` or something similar to group the more technical details together.

Finally, I encourage you to consider how permanent those hyperlinks will be. Consider that if the location or name of the markdown files are changed, the links will be broken. I would encourage you to include more of the relevant information to the paper in the paper itself, and allow the readme to link to the other documentation.

## Nitpicks
- Use the articles when referring to Jacobians. E.g. "The Jacobian" in the first two sentences.
- Move the GOGU2005427 citation to the end of the sentence.
- "all these studies `are dependent` on human inspection"
- "but either their application is limited to serial manipulators`,` they require human intervention`,` or they are part of computationally expensive simulations"
