+++ 4R4P.md
@@ -92,7 +92,7 @@ active_joint_velocities
 
 In an ipynb file of JupyterLab, the above code would produce the following output.
 
-$$\left[\begin{matrix}\dot{d}\_{(1,2)}\\\dot{d}\_{(1,3)}\\\dot{\theta}\_{(4,5)}\end{matrix}\right]$$
+$$\left[\begin{matrix}\dot{d}\_{(1,2)} \\\\ \dot{d}\_{(1,3)} \\\\ \dot{\theta}\_{(4,5)}\end{matrix}\right]$$
 
 Robot dimensional parameters can be viewed by running the below line.
 ```py
