+++ 4R4P.md
@@ -102,7 +102,7 @@ robot_dimensional_parameters
 
 In an ipynb file of JupyterLab, the above code would produce the following output.
 
-$$\left[\begin{matrix}r_{(1,2)x} \\\\ r_{(1,2)y} \\\\ r_{(2,3)x} \\\\ r_{(2,3)y}\end{matrix}\right]$$
+$$\left[\begin{matrix}\phi_{(1,2)}\\\\\phi_{(1,3)}\\\\\phi_{(2,4)}\\\\\phi_{(3,4)}\\\\r_{(4,5)x}\\\\r_{(4,5)y}\\\\r_{(4,6)x}\\\\r_{(4,6)y}\\\\r_{(5,7)x}\\\\r_{(5,7)y}\\\\r_{(6,7)x}\\\\r_{(6,7)y}\end{matrix}\right]$$
 
 
 Robot end-effector parameters can be viewed by running the below line.
