### 2023-11-03 65ee882
Update functions.py
Added `superfluous_information` attribute to `jacobian` class.
