### 2023-11-03 8a967a2
Update functions.py
Changed the attribute name
