+++ 4R4P.md
@@ -121,7 +121,7 @@ $$\left[\begin{matrix}a_{x} \\\\ a_{y}\end{matrix}\right]$$
 - Locations of joints: $\textbf{r}\_{(1,2)}=2\hat{i}+8\hat{j}$, $\textbf{r}\_{(2,3)}=6\hat{i}+8\hat{j}$, $\textbf{r}\_{(1,2)}=3\hat{i}+10\hat{j}$ and $\textbf{r}\_{(2,3)}=5\hat{i}+10\hat{j}$
 - Orientations of joints: $\phi\_{(1,2)}=2\pi/3$, $\phi\_{(1,3)}=\pi/3$, $\phi\_{(2,4)}=\pi/3$ and $\phi\_{(3,4)}=2\pi/3$
 
-For the given set of dimensional parameters of the robot, the numerical Jacobian can be computed as follows. Firstly, we need to gather the configuration parameters in Python list format, in a particular order. The robot dimensional parameters from `jac.parameters_symbolic` are found (as shown earlier) to be in the order of $r_{(1,2)x}$, $r_{(1,2)y}$, $r_{(2,3)x}$ and $r_{(2,3)y}$. Hence the configuration parameters are to be supplied in the same order, as a list. Thus, the computation can be performed as shown below.
+For the given set of dimensional parameters of the robot, the numerical Jacobian can be computed as follows. Firstly, we need to gather the configuration parameters in Python list format, in a particular order. The robot dimensional parameters from `jac.parameters_symbolic` are found (as shown earlier) to be in the order of $\phi_{(1,2)}$, $\phi_{(1,3)}$, $\phi_{(2,4)}$, $\phi_{(3,4)}$, $r_{(4,5)x}$, $r_{(4,5)y}$, $r_{(4,6)x}$, $r_{(4,6)y}$, $r_{(5,7)x}$, $r_{(5,7)y}$, $r_{(6,7)x}$ and $r_{(6,7)y}$. Hence the configuration parameters are to be supplied in the same order, as a list. Thus, the computation can be performed as shown below.
 ```py
 end_effector_point = [1,2]
 configuration_parameters = [3,4,5,6]
