+++ 4R4P.md
@@ -133,9 +133,9 @@ jacobian_at_the_given_configuration
 
 The output produced by running the above code, is shown below.
 ```py
-array([[ 2,  4],
-       [-2, -4],
-       [ 1,  1]])
+array([[ -0.5      ,   0.5      , -10.       ],
+       [  0.8660254,   0.8660254,   3.       ],
+       [  0.       ,   0.       ,  -1.       ]])
 ```
 
 ## Community Guidelines
