+++ 4R4P.md
@@ -138,11 +138,54 @@ array([[ -0.5      ,   0.5      , -10.       ],
        [  0.       ,   0.       ,  -1.       ]])
 ```
 
-## Community Guidelines
+#### Accessing each matrix individually:
 
-- For contribution to the software:
-  - In order to contribute to the software, please consider using the [pull request feature](https://github.com/suneeshjacob/ACRoD/issues) of GitHub.
-- For reporting issues with the software:
-  - For reporting issues or problems, please use [issues](https://github.com/suneeshjacob/ACRoD/issues).
-- For support:
-  - For any further support (including installation, usage, etc.), feel free to contact via suneeshjacob-at-gmail-dot-com.
+Each of $J_a$, $J_p$, $A_a$ and $A_p$ functions can be accessed as shown below
+
+```py
+numerical_Ja = jac.Ja_func(end_effector_point, configuration_parameters)
+numerical_Ja
+```
+Output:
+```py
+array([[-0.5      ,  0.       ,  6.       ],
+       [ 0.8660254,  0.       , -1.       ],
+       [ 0.       ,  0.       ,  1.       ]])
+```
+
+```py
+numerical_Jp = jac.Jp_func(end_effector_point, configuration_parameters)
+numerical_Jp
+```
+Output:
+```py
+array([[ 0.5      ,  0.       ,  0.       ,  8.       ,  0.       ],
+       [ 0.8660254,  0.       ,  0.       , -2.       ,  0.       ],
+       [ 0.       ,  0.       ,  0.       ,  1.       ,  0.       ]])
+```
+
+```py
+numerical_Aa = jac.Aa_func(end_effector_point, configuration_parameters)
+numerical_Aa
+```
+Output:
+```py
+array([[ 0.       ,  0.       , -6.       ],
+       [ 0.       ,  0.       ,  1.       ],
+       [ 0.5      ,  0.5      ,  0.       ],
+       [-0.8660254,  0.8660254,  0.       ],
+       [ 0.       ,  0.       , -1.       ]])
+```
+
+```py
+numerical_Ap = jac.Ap_func(end_effector_point, configuration_parameters)
+numerical_Ap
+```
+Output:
+```py
+array([[ 0.       ,  0.       ,  6.       , -8.       ,  8.       ],
+       [ 0.       ,  0.       , -5.       ,  2.       , -4.       ],
+       [-0.5      , -0.5      ,  0.       ,  0.       ,  0.       ],
+       [-0.8660254,  0.8660254,  0.       ,  0.       ,  0.       ],
+       [ 0.       ,  0.       ,  1.       , -1.       ,  1.       ]])
+```
