+++ 4R4P.md
@@ -189,3 +189,9 @@ array([[ 0.       ,  0.       ,  6.       , -8.       ,  8.       ],
        [-0.8660254,  0.8660254,  0.       ,  0.       ,  0.       ],
        [ 0.       ,  0.       ,  1.       , -1.       ,  1.       ]])
 ```
+
+And the computation $J_a-J_pA^{-1}_pAa$ is given by
+```py
+numerical_Ja-numpy.matmul(numpy.matmul(numerical_Jp,numpy.linalg.inv(numerical_Ap)),numerical_Aa)
+```
+which gives the same output as `jacobian_at_the_given_configuration`.
