+++ 4R4P.md
@@ -195,3 +195,17 @@ And the computation $J_a-J_pA^{-1}_pAa$ is given by
 numerical_Ja-numpy.matmul(numpy.matmul(numerical_Jp,numpy.linalg.inv(numerical_Ap)),numerical_Aa)
 ```
 which gives the same output as `jacobian_at_the_given_configuration`.
+
+
+### Some other attributes
+
+To get the computed list of all connecting paths from the base link to the end-effector link, the below script can be used:
+
+```py
+jac.P
+```
+Output:
+```py
+[[0, 1, 3, 4, 6], [0, 1, 3, 5, 6], [0, 2, 3, 4, 6], [0, 2, 3, 5, 6]]
+```
+which gives the list of all connecting paths (only link numbers are shown, indexed from 0).
