+++ 4R4P.md
@@ -209,3 +209,25 @@ Output:
 [[0, 1, 3, 4, 6], [0, 1, 3, 5, 6], [0, 2, 3, 4, 6], [0, 2, 3, 5, 6]]
 ```
 which gives the list of all connecting paths (only link numbers are shown, indexed from 0).
+
+Out of these paths, the computed list of independent paths pertaining to linear velocities can be accessed by the below script:
+
+```py
+jac.P_tilde
+```
+Output:
+```py
+[[0, 1, 3, 4, 6], [0, 1, 3, 5, 6], [0, 2, 3, 4, 6]]
+```
+which gives the list of all the independent connecting paths pertaining to linear velocities (only link numbers are shown, indexed from 0)
+
+and the computed list of independent paths pertaining to angular velocities can be accessed by the below script:
+
+```py
+jac.P_tilde_omega
+```
+Output:
+```py
+[[0, 1, 3, 4, 6], [0, 1, 3, 5, 6]]
+```
+which gives the list of all the independent connecting paths pertaining to angular velocities (only link numbers are shown, indexed from 0).
