+++ 4R4P.md
@@ -212,8 +212,9 @@ which gives the list of all connecting paths (only link numbers are shown, index
 
 #### Independent paths pertaining to linear and angular velocities:
 
-Out of the above paths, the computed list of independent paths pertaining to linear velocities can be accessed by the below script:
+Out of the above paths, the computed list of independent paths pertaining to linear velocities and angular velocities, can be accessed by the below scripts:
 
+For independent paths pertaining to linear velocities:
 ```py
 jac.P_tilde
 ```
@@ -221,10 +222,7 @@ Output:
 ```py
 [[0, 1, 3, 4, 6], [0, 1, 3, 5, 6], [0, 2, 3, 4, 6]]
 ```
-which gives the list of all the independent connecting paths pertaining to linear velocities (only link numbers are shown, indexed from 0),
-
-and the computed list of independent paths pertaining to angular velocities can be accessed by the below script:
-
+For independent paths pertaining to angular velocities:
 ```py
 jac.P_tilde_omega
 ```
@@ -232,4 +230,4 @@ Output:
 ```py
 [[0, 1, 3, 4, 6], [0, 1, 3, 5, 6]]
 ```
-which gives the list of all the independent connecting paths pertaining to angular velocities (only link numbers are shown, indexed from 0).
+The above scripts give the lists of all the independent connecting paths pertaining to linear and angular velocities (only link numbers are shown, indexed from 0).
