### 2023-11-04 3a14362
Create RSSRSSR.md
Added RSSRSSR robot example
