+++ RSSRSSR.md
@@ -1,0 +1,233 @@
+# Formulation of Jacobian for RSSR-SSR spatial parallel manipulator (with a superfluous ODF) as an example
+
+## Figure
+
+<p align="center">
+    <img src="./examples/Jacobian/RSSRSSR.png" alt="RSSR-SSR spatial parallel manipulator" width="500px">
+</p>
+
+A figure of RRRRPPPP planar serial-parallel hybrid manipulator is shown above. The corresponding adjacency matrix is given by
+
+$$\bf{M} = \left[\begin{matrix}L_1 & P & P & O & O & O & O\\\\A & L_2 & O & P & O & O & O\\\\A & O & L_3 & P & O & O & O\\\\O & O & O & L_4 & R & R & O\\\\O & O & O & A & L_5 & O & R\\\\O & O & O & O & O & L_6 & R\\\\O & O & O & O & O & O & L_7\end{matrix}\right]$$
+
+## Usage
+
+### Jacobian for planar manipulators
+
+The topological information of a robot is to be specified by using its robot-topology matrix, as defined [here](Robot_Topology_Matrix.md). For RRRRPPPP planar serial-parallel hybrid manipulator shown above, the robot topology matrix is given by
+
+$$\left[\begin{matrix}9 & 2 & 2 & 0 & 0 & 0 & 0\\\\1 & 9 & 0 & 2 & 0 & 0 & 0\\\\1 & 0 & 9 & 2 & 0 & 0 & 0\\\\0 & 0 & 0 & 9 & 1 & 1 & 0\\\\0 & 0 & 0 & 1 & 9 & 0 & 1\\\\0 & 0 & 0 & 0 & 0 & 9 & 1\\\\0 & 0 & 0 & 0 & 0 & 0 & 9\end{matrix}\right]$$
+
+The corresponding Jacobian function can be formulated as follows.
+
+Firstly, the required functions are imported as shown below.
+```py
+from acrod.functions import jacobian
+from numpy import array
+```
+
+
+The robot-topology matrix for 3R planar serial manipulator is defined and jacobian information is processed via the imported jacobian class as follows.
+```py
+M = numpy.array(
+        [[9, 2, 2, 0, 0, 0, 0],
+         [1, 9, 0, 2, 0, 0, 0],
+         [1, 0, 9, 2, 0, 0, 0],
+         [0, 0, 0, 9, 1, 1, 0],
+         [0, 0, 0, 1, 9, 0, 1],
+         [0, 0, 0, 0, 0, 9, 1],
+         [0, 0, 0, 0, 0, 0, 9]]
+    )
+jac = jacobian(M, robot_type = 'planar')
+```
+
+
+Jacobian function is generated as shown below.
+```py
+jacobian_function = jac.get_jacobian_function()
+```
+
+
+In the process of generating the above jacobian function, other attributes of the jacobian object also are updated. Symbolic Jacobian matrices can be extracted from the attributes. Since this is a non-serial robot, there would be four matrices required to compute the Jacobian, which are $J_a$, $J_p$, $A_a$ and $A_p$. These can be extracted by the attributes `Ja`, `Jp`, `Aa` and `Ap`, respectively, as shown below.
+
+```py
+symbolic_Ja = jac.Ja
+symbolic_Ja
+```
+
+Output in Jupyter notebook:
+
+$$\left[\begin{matrix}\cos{\left(\phi_{(1,2)} \right)} & 0 & - a_{y} + r_{(4,5)y}\\\\\sin{\left(\phi_{(1,2)} \right)} & 0 & a_{x} - r_{(4,5)x}\\\\0 & 0 & 1\end{matrix}\right]$$
+
+```py
[…]
