+++ 4R4P.md
@@ -190,7 +190,7 @@ array([[ 0.       ,  0.       ,  6.       , -8.       ,  8.       ],
        [ 0.       ,  0.       ,  1.       , -1.       ,  1.       ]])
 ```
 
-And the computation $J_a-J_pA^{-1}_pAa$ is given by
+And the computation $J_a-J_pA^{-1}_pA_a$ is given by
 ```py
 numerical_Ja-numpy.matmul(numpy.matmul(numerical_Jp,numpy.linalg.inv(numerical_Ap)),numerical_Aa)
 ```
