+++ RSSRSSR.md
@@ -85,7 +85,7 @@ symbolic_Ap
 ```
 Output in Jupyter notebook:
 
-$$\left[\begin{array}{ccccccccccccc}- \left(a_{y} - r_{(1,5)y}\right) \cos{\left(\beta_{(1,5)} \right)} + \left(a_{z} - r_{(1,5)z}\right) \sin{\left(\beta_{(1,5)} \right)} \sin{\left(\phi_{(1,5)} \right)} & 0 & - a_{z} + r_{(2,6)z} & a_{y} - r_{(2,6)y} & 0 & 0 & 0 & 0 & 0 & 0 & 0 & a_{z} - r_{(5,6)z} & - a_{y} + r_{(5,6)y}\\\\\left(a_{x} - r_{(1,5)x}\right) \cos{\left(\beta_{(1,5)} \right)} - \left(a_{z} - r_{(1,5)z}\right) \sin{\left(\beta_{(1,5)} \right)} \cos{\left(\phi_{(1,5)} \right)} & a_{z} - r_{(2,6)z} & 0 & - a_{x} + r_{(2,6)x} & 0 & 0 & 0 & 0 & 0 & 0 & - a_{z} + r_{(5,6)z} & 0 & a_{x} - r_{(5,6)x}\\\\- \left(a_{x} - r_{(1,5)x}\right) \sin{\left(\beta_{(1,5)} \right)} \sin{\left(\phi_{(1,5)} \right)} + \left(a_{y} - r_{(1,5)y}\right) \sin{\left(\beta_{(1,5)} \right)} \cos{\left(\phi_{(1,5)} \right)} & - a_{y} + r_{(2,6)y} & a_{x} - r_{(2,6)x} & 0 & 0 & 0 & 0 & 0 & 0 & 0 & a_{y} - r_{(5,6)y} & - a_{x} + r_{(5,6)x} & 0\\\\0 & 0 & - a_{z} + r_{(2,6)z} & a_{y} - r_{(2,6)y} & 0 & a_{z} - r_{(3,4)z} & - a_{y} + r_{(3,4)y} & 0 & a_{z} - r_{(4,6)z} & - a_{y} + r_{(4,6)y} & 0 & 0 & 0\\\\0 & a_{z} - r_{(2,6)z} & 0 & - a_{x} + r_{(2,6)x} & - a_{z} + r_{(3,4)z} & 0 & a_{x} - r_{(3,4)x} & - a_{z} + r_{(4,6)z} & 0 & a_{x} - r_{(4,6)x} & 0 & 0 & 0\\\\0 & - a_{y} + r_{(2,6)y} & a_{x} - r_{(2,6)x} & 0 & a_{y} - r_{(3,4)y} & - a_{x} + r_{(3,4)x} & 0 & a_{y} - r_{(4,6)y} & - a_{x} + r_{(4,6)x} & 0 & 0 & 0 & 0\\\\\sin{\left(\beta_{(1,5)} \right)} \cos{\left(\phi_{(1,5)} \right)} & -1 & 0 & 0 & 0 & 0 & 0 & 0 & 0 & 0 & 1 & 0 & 0\\\\\sin{\left(\beta_{(1,5)} \right)} \sin{\left(\phi_{(1,5)} \right)} & 0 & -1 & 0 & 0 & 0 & 0 & 0 & 0 & 0 & 0 & 1 & 0\\\\\cos{\left(\beta_{(1,5)} \right)} & 0 & 0 & -1 & 0 & 0 & 0 & 0 & 0 & 0 & 0 & 0 & 1\\\\0 & -1 & 0 & 0 & 1 & 0 & 0 & 1 &
[…]
