### 2023-11-06 #5927 @gfadini
## Summary
The main stated contribution of this library is to provide a framework for the automatic calculation of the Jacobian in the case of several robot topologies. The aim is to simplify and automatize the tedious computations to obtain the jacobian of the end-effector. The authors present the procedure in `algorithm1.png` for this computation, based on an augmented adjacency matrix defining the robot topology.

## Main comments
Several points need to be addressed in order to increase the quality of the contribution:
- How does your library place with respect to current solutions?
	1) Several libraries implement algorithms to assess performance metrics of the end-effector (jacobian is just an intermediate step). For instance `PyCapacity`, another JOSS contribution, aims at solving a similar problem by the use of polytopes and ellipsoids. To this end, I was expecting the automatic computation of at least a few manipulability indicators (e.g. eigenvalues, matrix norms, condition number, et cetera).
	2) Several robotic libraries (e.g. `pinocchio`) compute the jacobian of the end-effector and can be easily used to get closed-loop jacobians. In the case of closed-loop kinematic chains, if the system is not degenerate, by the use of the implicit function theorem, it is possible to express the passive velocities as a function of the independent ones. This approach is detailed, for instance in [this paper](https://hal.science/hal-04191553/document), which also introduces useful performance metrics.
- The statement that the design of a robot can be automatized by the use of this library needs to be proven with examples.  Also the title of the paper "Automatic computation of robot design" has little to do with the contribution scope, and this may be confusing for the potential users of the tool. At the current stage, it seems that no form of automatic design generation or optimization (this can be quickly checked as no optimization library is required).
- The methodology heavily uses a custom encoding of the topology explained in the readme `Robot_Topology_Matrix`. This mathematical object is misleadingly called an adjacency matrix (as, following the reasoning, it may even be nonsymmetric). In reality, the described matrix is a custom encoding of the robot topology and joint types (where upper and lower diagonal elements encode the joint type and if it is associated with an independent variable). For more complex robots it may become very error-prone, and it seems to me that no checks on the matrix (see following comments) are made before manipulating it. Would the library benefit from the use of more standardized formats like `urdf` or `srdf` instead?

## Code/implementation details
On the library:
	- the code generally has comments, but the repo is missing examples.
	- more modularity in the code structure would be appreciated, for now, there is just a generic `functions.py` which provides all the repo functionality.
	- notes on `Robot_Topology_Matrix.md`:
		- Repetition of "The adjacency matrix is normally a symmetric matrix"
		- What happens if the system has multiple end effectors? Do you need to permutate the matrix at each time?
		- How do you treat the constraints arising from between active passive joints? To be more clear, once you have a kinematic loop, not all joints can be active, and the difference between active and passive joints needs to equal the free DoF of the chain.
		- When explaining the meaning of the lower diagonal part of the matrix I see a contradiction. On one side the goal of the library is to allow the automatic computation of the jacobian, but as pointed out there is some freedom in the specification of the lower part. This is indeed a cumbersome step that needs additional insights.

## Testing
Unfortunately, I could not make use of the library in some simple cases. The example runs but testing the library for a serial revolute joint manipulator does not give the expected outputs. Moreover, it seems that not all the joint types are functional. See the driving code [here](https://gist.github.com/gfadini/6a99978064fc51b24d0327e84fcc4dec)
