### 2023-11-07 301422a
Update functions.py
Fixed a bug pointed out by the reviewer for JOSS: `gfadini`
