### 2023-11-07 396a9bf
Update functions.py
Fixed a bug
