+++ RSSRSSR.md
@@ -129,9 +129,12 @@ jacobian_at_the_given_configuration
 
 The output produced by running the above code, is shown below.
 ```py
-array([[ -0.5      ,   0.5      , -10.       ],
-       [  0.8660254,   0.8660254,   3.       ],
-       [  0.       ,   0.       ,  -1.       ]])
+array([[ 19.62196517,  10.77434097],
+       [-16.1460825 ,  -8.61947278],
+       [ 43.06469529,  30.16815472],
+       [  6.57574977,   4.30973639],
+       [ 19.33597911,  12.92920917],
+       [  5.34317907,   2.15486819]])
 ```
 
 #### Accessing each matrix individually:
@@ -144,9 +147,12 @@ numerical_Ja
 ```
 Output:
 ```py
-array([[-0.5      ,  0.       ,  6.       ],
-       [ 0.8660254,  0.       , -1.       ],
-       [ 0.       ,  0.       ,  1.       ]])
+array([[ 5.19615242,  0.        ],
+       [-0.8660254 ,  0.        ],
+       [-1.0669873 ,  0.        ],
+       [ 0.25      ,  0.        ],
+       [ 0.4330127 ,  0.        ],
+       [ 0.8660254 ,  0.        ]])
 ```
 
 ```py
@@ -155,9 +161,12 @@ numerical_Jp
 ```
 Output:
 ```py
-array([[ 0.5      ,  0.       ,  0.       ,  8.       ,  0.       ],
-       [ 0.8660254,  0.       ,  0.       , -2.       ,  0.       ],
-       [ 0.       ,  0.       ,  0.       ,  1.       ,  0.       ]])
+array([[ 0,  0,  1, -1,  0,  0,  0,  0,  0,  0,  0,  0,  0],
+       [ 0, -1,  0, -2,  0,  0,  0,  0,  0,  0,  0,  0,  0],
+       [ 0,  1,  2,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0],
+       [ 0,  1,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0],
+       [ 0,  0,  1,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0],
+       [ 0,  0,  0,  1,  0,  0,  0,  0,  0,  0,  0,  0,  0]])
 ```
 
 ```py
@@ -166,24 +175,40 @@ numerical_Aa
 ```
 Output:
 ```py
-array([[ 0.       ,  0.       , -6.       ],
-       [ 0.       ,  0.       ,  1.       ],
-       [ 0.5      ,  0.5      ,  0.       ],
-       [-0.8660254,  0.8660254,  0.       ],
-       [ 0.       ,  0.       , -1.       ]])
+array([[-5.19615242,  0.        ],
+       [ 0.8660254 ,  0.        ],
+       [ 1.0669873 ,  0.        ],
+       [-5.19615242, -0.8660254 ],
+       [ 0.8660254 , -0.5       ],
+       [ 1.0669873 , -0.        ],
+       [-0.25      ,  0.        ],
+       [-0.4330127 ,  0.        ],
+       [-0.8660254 ,  0.        ],
+       [-0.25      , -0.25      ],
+       [-0.4330127 ,  0.4330127 ],
+       [-0.8660254 , -0.8660254 ],
+       [ 0.        ,  2.66506351]])
 ```
 
 ```py
[…]
