### 2023-11-08 3e79106
Update functions.py
Changed `self.superfluous_information` to `self.superfluous_dof_information`
