+++ RRR.md
@@ -98,168 +98,16 @@ For the given set of dimensional parameters of the robot, the numerical Jacobian
 from numpy import pi
 
 end_effector_point = [1,2]
-configuration_parameters = [2, 8, 3, pi/6, pi/3, 1, 2, 5, 5*pi/6, 2*pi/3, 2, 4, 7, pi/12, pi/2, 3, 1, 2, 6, 8, 4, 8, 1, 3, 5, 7, 3]
+configuration_parameters = [3, 4, 5, 6, 2, 3]
 jacobian_at_the_given_configuration = jacobian_function(end_effector_point, configuration_parameters)
 jacobian_at_the_given_configuration
 ```
 
 The output produced by running the above code, is shown below.
 ```py
-array([[ 19.62196517,  10.77434097],
-       [-16.1460825 ,  -8.61947278],
-       [ 43.06469529,  30.16815472],
-       [  6.57574977,   4.30973639],
-       [ 19.33597911,  12.92920917],
-       [  5.34317907,   2.15486819]])
+array([[ 2,  4,  1],
+       [-2, -4, -1],
+       [ 1,  1,  1]])
 ```
 
-#### Accessing each matrix individually:
 
-Each of $J_a$, $J_p$, $A_a$ and $A_p$ functions can be accessed as shown below
-
-```py
-numerical_Ja = jac.Ja_func(end_effector_point, configuration_parameters)
-numerical_Ja
-```
-Output:
-```py
-array([[ 5.19615242,  0.        ],
-       [-0.8660254 ,  0.        ],
-       [-1.0669873 ,  0.        ],
-       [ 0.25      ,  0.        ],
-       [ 0.4330127 ,  0.        ],
-       [ 0.8660254 ,  0.        ]])
-```
-
-```py
-numerical_Jp = jac.Jp_func(end_effector_point, configuration_parameters)
-numerical_Jp
-```
-Output:
-```py
-array([[ 0,  0,  1, -1,  0,  0,  0,  0,  0,  0,  0,  0,  0],
-       [ 0, -1,  0, -2,  0,  0,  0,  0,  0,  0,  0,  0,  0],
-       [ 0,  1,  2,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0],
-       [ 0,  1,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0],
-       [ 0,  0,  1,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0],
-       [ 0,  0,  0,  1,  0,  0,  0,  0,  0,  0,  0,  0,  0]])
-```
-
-```py
-numerical_Aa = jac.Aa_func(end_effector_point, configuration_parameters)
-numerical_Aa
-```
-Output:
-```py
-array([[-5.19615242,  0.        ],
-       [ 0.8660254 ,  0.        ],
-       [ 1.0669873 ,  0.        ],
-       [-5.19615242, -0.8660254 ],
-       [ 0.8660254 , -0.5       ],
-       [ 1.0669873 , -0.        ],
-       [-0.25      ,  0.        ],
-       [-0.4330127 ,  0.        ],
-       [-0.8660254 ,  0.        ],
-       [-0.25      , -0.25      ],
-       [-0.4330127 ,  0.4330127 ],
-       [-0.8660254 , -0.8660254 ],
-       [ 0.        ,  2.66506351]])
-```
-
-```py
-numerical_Ap = jac.Ap_func(end_effector_point, configuration_parameters)
-numerical_Ap.round()
-```
-Output:
-```py
[…]
