+++ RRR.md
@@ -14,7 +14,7 @@ $$\bf{M} = \left[\begin{matrix}L_1 & R & O & O \\\\A & L_2 & R & O\\\\O & A & L_
 
 ### Jacobian for planar manipulators
 
-The topological information of a robot is to be specified by using its robot-topology matrix, as defined [here](Robot_Topology_Matrix.md). For RRRRPPPP planar serial-parallel hybrid manipulator shown above, the robot topology matrix is given by
+The topological information of a robot is to be specified by using its robot-topology matrix, as defined [here](Robot_Topology_Matrix.md). For RRR planar serial manipulator shown above, the robot topology matrix is given by
 
 $$\left[\begin{matrix}9 & 1 & 0 & 0 \\\\ 1 & 9 & 1 & 0 \\\\ 0 & 1 & 9 & 1 \\\\ 0 & 0 & 1 & 9\end{matrix}\right]$$
 
