### 2023-11-09 bb40098
Update 3R_serial_robot.md
