+++ examples/Jacobian/codes/4R4P_code.md
@@ -3,7 +3,7 @@
 ## Figure
 
 <p align="center">
-    <img src="./examples/Jacobian/RRRRPPPP.png" alt="4R-4P planar serial-parallel hybrid manipulator" width="500px">
+    <img src="../images/RRRRPPPP.png" alt="4R-4P planar serial-parallel hybrid manipulator" width="500px">
 </p>
 
 A figure of RRRRPPPP planar serial-parallel hybrid manipulator is shown above. The corresponding adjacency matrix is given by
