### 2023-11-09 78efd64
Update 3R_serial_robot.md
