+++ examples/Jacobian/maths/3R_serial_robot.md
@@ -3,7 +3,7 @@
 ## Mathematics involved
 
 <p align="center">
-    <img src="./examples/Jacobian/images/RRR.png" alt="RRR planar serial manipulator" width="500px">
+    <img src="../images/RRR.png" alt="RRR planar serial manipulator" width="500px">
 </p>
 
 An RRR planar serial manipulator is considered as shown in the figure \ref{fig:RRR}. The corresponding adjacency matrix is given by
