### 2023-11-09 400f8b0
Update 4R4P_parallel_robot.md
