### 2023-11-09 79a921f
Update RSSR-SSR_serial-parallel_hybrid_robot copy.md
