### 2023-11-09 79510d4
Update RSSR-SSR_serial-parallel_hybrid_robot.md
