### 2023-11-09 de93034
Update 3R_serial_robot.md
