### 2023-11-09 c4cecf5
Update 3R_serial_robot.md
