### 2023-11-09 313ea27
Update 3R_serial_robot.md
