### 2023-11-09 1ee292e
Update 4R4P_parallel_robot.md
