### 2023-11-09 f4a8fe2
Update RSSR-SSR_serial-parallel_hybrid_robot.md
