### 2023-11-09 a86e735
Update RSSR-SSR_serial-parallel_hybrid_robot.md
