### 2023-11-09 bbd787a
Update RSSR-SSR_serial-parallel_hybrid_robot.md
