### 2023-11-09 7b302e4
Update RSSR-SSR_serial-parallel_hybrid_robot.md
