+++ examples/Jacobian/maths/RSSR-SSR_serial-parallel_hybrid_robot.md
@@ -153,7 +153,7 @@ $$
 
 The system of equations shown above would be insufficient to represent the passive joint velocities in terms of active joint velocities. This is due to the fact that the fourth link has rotation along its longitudinal axis not controllable with the actuators yet does not affect the end-effector's velocity. Performing pseudo-inverse can fix this issue but pseudo-inverse could be a discontinuous operation near singular values. Instead, an additional equation for each such superfluous DOF is calculated as per the method shown in algorithm 3 of the main document.
 
-Since the manipulator has more than two spherical joints, the list of all possible combinations C of dividing the manipulator into two parts is considered. Since there are six links and since 6 is an odd number, ${}^{6}C_{1}+{}^{6}C_{2}+{}^{6}C_{3}=41$ distinct combinations exist, out of which the combination $[\{4\}, \{1,2,3,5,6\}]$ is discussed in detail in this sub-section. This combination has $c_1 = \\{4\\}$ and $c_2 = \\{1,2,3,5,6\\}$. And by grouping the links of each part together, the topology-matrix can be rewritten as shown in the equation below.
+Since the manipulator has more than two spherical joints, the list of all possible combinations C of dividing the manipulator into two parts is considered. Since there are six links and since 6 is an odd number, ${}^{6}C_{1}+{}^{6}C_{2}+{}^{6}C_{3}=41$ distinct combinations exist, out of which the combination $[\\{4\\}, \\{1,2,3,5,6\\}]$ is discussed in detail in this sub-section. This combination has $c_1 = \\{4\\}$ and $c_2 = \\{1,2,3,5,6\\}$. And by grouping the links of each part together, the topology-matrix can be rewritten as shown in the equation below.
 
 $$\widetilde{M} = \begin{bmatrix}
     L_4 & O & O & S & O & S \\\\
