+++ paper/paper.md
@@ -33,7 +33,7 @@ Jacobian of a robot refers to the matrix that linearly maps the velocity compone
 
 For a manipulator of a given topology, designing the dimensions based on optimising Jacobian-based performance parameters (such as manipulability index and condition number [@patel2015manipulator]) around a given end-effector point would require only the topological information for the formulation of Jacobian, as every other step can be automated. Formulation of Jacobian for parallel manipulators and serial-parallel hybrid manipulators are non-trivial, although all the steps of Jacobian formulation even in those cases would have to stem from the mere information of topology of the robot. ACRoD automates the non-trivial formulation of Jacobian systematically. It uses a matrix-based representation of the topology of the robotic manipulator (referred to here as the robot-topology matrix, of which more information is provided [here](https://github.com/suneeshjacob/ACRoD/blob/main/Robot_Topology_Matrix.md)) which is a modified version of the graph adjacency matrix representation [@enumeration] of robotic manipulators. This Jacobian formulation can be used to generate numerical Jacobian matrices with a few random configurations, from which the singular values can be calculated which can confirm the Degree of Freedom (DoF). In other words, the DoF of a given robot topology for a given base link and a given end-effector link can be verified by using this Jacobian function even in cases where Chebychev–Grübler–Kutzbach criterion fails to verify. This can be useful in mechanism synthesis to accurately verify the mobility of a given manipulator directly from its robot-topology matrix by using the method shown in Yang et al.'s paper [@YANG20081175].
 
-`ACRoD` uses NumPy [@numpy] and SymPy [@sympy] packages to generate the functions for Jacobian, which can be directly used in optimisation process to find the optimal dimensional parameters of the robot. The topology of a valid robot (with a single base-link and a single end-effector link and without non-contributing chains) is to be specified using robot-topology matrix in NumPy matrix format. The jacobian class object takes this robot-topology matrix as input argument and generates functions that are required to compute Jacobian. As byproducts, the Jacobian function generation produces the symbolic matrices, the set of independent paths, etc., the sets of active joint velocities and passive joint velocities, etc., which can be accessed from the attributes of the jacobian class object. More technical details on formulation of Jacobian (along with appropriate algorithms) can be found [here](https://github.com/suneeshjacob/ACRoD/blob/main/Mathematics%20behind%20Jacobian%20formulation.md), and the notations and the nomenclature are explained [here](https://github.com/suneeshjacob/ACRoD/blob/main/Notation_and_Nomenclature.md) in detail. The robot-topology matrix representation is explained [here](https://github.com/suneeshjacob/ACRoD/blob/main/Robot_Topology_Matrix.md) in detail. Jacobian formulation for three robot examples, namely the [3R planar serial robot](https://github.com/suneeshjacob/ACRoD/blob/main/examples/Jacobian/maths/3R_serial_robot.md), [a 4R-4P planar serial-parallel hybrid robot](https://github.com/suneeshjacob/ACRoD/blob/main/examples/Jacobian/maths/4R4P_parallel_robot.md) and [an RSSR-SSR spatial parallel robot](https://github.com/suneeshjacob/ACRoD/blob/main/examples/Jacobian/maths/RSSR-SSR_serial-parallel_hybrid_robot%20copy.md), are explained in detail in the corresponding hyperlinks.
[…]
