+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "acrod"
-version = "1.1.4"
+version = "1.1.5"
 authors = [
   { name="Suneesh Jacob Akkarapakam", email="suneeshjacob@gmail.com" },
   { name="Rituparna Datta", email="rituparndatta@gmail.com" }
