+++ examples/Jacobian/codes/4R4P_code.md
@@ -29,7 +29,7 @@ from numpy import array
 
 The robot-topology matrix for 3R planar serial manipulator is defined and jacobian information is processed via the imported jacobian class as follows.
 ```py
-M = numpy.array(
+M = array(
         [[9, 2, 2, 0, 0, 0, 0],
          [1, 9, 0, 2, 0, 0, 0],
          [1, 0, 9, 2, 0, 0, 0],
