+++ paper/paper.md
@@ -39,11 +39,13 @@ For a manipulator of a given topology, designing the dimensions based on optimis
 | Software    | Base Language   | Closed-loop linkages | Input complexity | Primary purpose       |
 |-------------|:---------------:|:--------------------:|------------------|-----------------------|
 | TriP (2022) | Python          | yes                  | non-strict       | Kinematics            |
-| Pybotics    | Python          | yes                  | strict           | 1993                  |
+| Pybotics    | Python          | no                   | strict           | 1993                  |
 | DART        | C++             | no                   | strict           | 1972                  |
 | CoreRobotics| C++             | no                   | strict           | 1972                  |
-| pinocchio   | C++             | no                   | strict           | Jacobian computation  |
+| pinocchio   | C++             | X                    | strict           | Jacobian computation  |
 | ACRoD       | Python          | yes                  | Matrix           | Dimensional Synthesis |
+
+X = possible but documentation not provided
 
 # Acknowledgements
 
