+++ paper/paper.bib
@@ -203,3 +203,27 @@ doi = {10.1016/j.mechmachtheory.2007.08.001}
   url = {https://doi.org/10.7717/peerj-cs.103},
   doi = {10.7717/peerj-cs.103}
 }
+
+@inproceedings{carpentier:hal-01866228,
+  TITLE = {{The Pinocchio C++ library -- A fast and flexible implementation of rigid body dynamics algorithms and their analytical derivatives}},
+  AUTHOR = {Carpentier, Justin and Saurel, Guilhem and Buondonno, Gabriele and Mirabel, Joseph and Lamiraux, Florent and Stasse, Olivier and Mansard, Nicolas},
+  URL = {https://laas.hal.science/hal-01866228},
+  BOOKTITLE = {{SII 2019 - International Symposium on System Integrations}},
+  ADDRESS = {Paris, France},
+  HAL_LOCAL_REFERENCE = {Rapport LAAS n{\textdegree} 18288},
+  YEAR = {2019},
+  MONTH = Jan,
+  PDF = {https://laas.hal.science/hal-01866228v2/file/19-sii-pinocchio.pdf},
+  HAL_ID = {hal-01866228},
+  HAL_VERSION = {v2},
+}
+
+@INPROCEEDINGS{10260660,
+  author={Tola, Daniella and Corke, Peter},
+  booktitle={2023 IEEE 19th International Conference on Automation Science and Engineering (CASE)}, 
+  title={Understanding URDF: A Survey Based on User Experience}, 
+  year={2023},
+  volume={},
+  number={},
+  pages={1-7},
+  doi={10.1109/CASE56687.2023.10260660}}
