+++ paper/paper.md
@@ -36,17 +36,18 @@ For a manipulator of a given topology, designing the dimensions based on optimis
 `ACRoD` uses NumPy [@numpy] and SymPy [@sympy] packages to generate the functions for Jacobian, which can be directly used in optimisation process to find the optimal dimensional parameters of the robot. The topology of a valid robot (with a single base-link and a single end-effector link and without non-contributing chains) is to be specified using robot-topology matrix in NumPy matrix format. The jacobian class object takes this robot-topology matrix as input argument and generates functions that are required to compute Jacobian. As byproducts, the Jacobian function generation produces the symbolic matrices, the set of independent paths, etc., the sets of active joint velocities and passive joint velocities, etc., which can be accessed from the attributes of the jacobian class object. More technical details on formulation of Jacobian (along with appropriate algorithms) can be found [here](https://github.com/suneeshjacob/ACRoD/blob/main/Mathematics%20behind%20Jacobian%20formulation.md), and the notations and the nomenclature are explained [here](https://github.com/suneeshjacob/ACRoD/blob/main/Notation_and_Nomenclature.md) in detail. The robot-topology matrix representation is explained [here](https://github.com/suneeshjacob/ACRoD/blob/main/Robot_Topology_Matrix.md) in detail. Jacobian formulation for three robot examples, namely the [3R planar serial robot](https://github.com/suneeshjacob/ACRoD/blob/main/examples/Jacobian/maths/3R_serial_robot.md), [a 4R-4P planar serial-parallel hybrid robot](https://github.com/suneeshjacob/ACRoD/blob/main/examples/Jacobian/maths/4R4P_parallel_robot.md) and [an RSSR-SSR spatial parallel robot](https://github.com/suneeshjacob/ACRoD/blob/main/examples/Jacobian/maths/RSSR-SSR_serial-parallel_hybrid_robot.md), are explained in detail in the corresponding hyperlinks.
 
 ## Comparison with other Jacobian-computation softwares
+TriP [@Baumgärtner2022] is a software that is developed to address kinematics of hybrid linkages. However, it is evident from the triped example that human inspection is required to develop the model for each leg of triped robot. DART [@Lee2018] uses biped robot (which can be seen as involving a closed-loop mechanism) in the examples, however it is imported from .sktl file rather than modelling a customised closed-loop robot from scratch. Even though C++ based softwares such as DART [@Lee2018], CoreRobotics [@Owan2018] and pinocchio [@carpentier:hal-01866228] may facilitate closed-loop linkages, apparently no documentation is provided for modelling closed-loop linkages. Furthermore, all these software (TriP, DART, CoreRobotics, pinocchio) apparently require human intervention to build a robot of a given topology. Even thogh some of them support importing models from URDF files, URDF has the limitation of "inability to model parallel linkages and closed-chain systems" [@10260660], and furthermore preparing a URDF file of a robot from its mere topological information also requires human intervention. Pybotics provides automatic modelling of robot from the mere information of DH parameters, however it is limited to serial manipulators. ACRoD addresses this issue of human intervention by automatically generating functions required to compute Jacobian for a given end-effector point. A comparison of ACRoD with other softwares is shown in the table below.
 | Software           | Base Language   | Closed-loop Linkages | Automation Level From Mere Topology        | Primary focus                                                                                      |
 |--------------------|:---------------:|:--------------------:|--------------------------------------------|----------------------------------------------------------------------------------------------------|
 | TriP (2022)        | Python          | yes                  | H*                                         | Kinematics of Hybrid Linkages                                                                      |
-| Pybotics (2019)    | Python          | no                   | A                                          |   Kinematics, Dynamics, Trajectory Generations and Calibration of Serial Robots                    |
[…]
