+++ paper/paper.md
@@ -37,6 +37,7 @@ For a manipulator of a given topology, designing the dimensions based on optimis
 
 ## Comparison with other Jacobian-computation softwares
 TriP [@Baumgärtner2022] is a software that is developed to address kinematics of hybrid linkages. However, it is evident from the triped example that human inspection is required to develop the model for each leg of triped robot. DART [@Lee2018] uses biped robot (which can be seen as involving a closed-loop mechanism) in the examples, however it is imported from .sktl file rather than modelling a customised closed-loop robot from scratch. Even though C++ based softwares such as DART [@Lee2018], CoreRobotics [@Owan2018] and pinocchio [@carpentier:hal-01866228] may facilitate closed-loop linkages, apparently no documentation is provided for modelling closed-loop linkages. Furthermore, all these software (TriP, DART, CoreRobotics, pinocchio) apparently require human intervention to build a robot of a given topology. Even thogh some of them support importing models from URDF files, URDF has the limitation of "inability to model parallel linkages and closed-chain systems" [@10260660], and furthermore preparing a URDF file of a robot from its mere topological information also requires human intervention. Pybotics provides automatic modelling of robot from the mere information of DH parameters, however it is limited to serial manipulators. ACRoD addresses this issue of human intervention by automatically generating functions required to compute Jacobian for a given end-effector point. A comparison of ACRoD with other softwares is shown in the table below.
+
 | Software           | Base Language   | Closed-loop Linkages | Automation Level From Mere Topology        | Primary focus                                                                                      |
 |--------------------|:---------------:|:--------------------:|--------------------------------------------|----------------------------------------------------------------------------------------------------|
 | TriP (2022)        | Python          | yes                  | H*                                         | Kinematics of Hybrid Linkages                                                                      |
