### 2023-12-01 e55dae5
Update paper.md
Made changes based on suggestions of the review by `JHartzer`, and made some other change(s).
