+++ paper/paper.md
@@ -27,7 +27,7 @@ bibliography: paper.bib
 
 # Summary
 
-Jacobian of a robot refers to the matrix that linearly maps the velocity components of the end-effector and the velocities at the actuated joints. Jacobian is extensively used in dimensional synthesis for Jacobian-based optimal performances of robotic manipulators, in which the optimal dimensional parameters of robots are computed. Determination of accurate mobility [@YANG20081175] of planar and spatial mechanisms can also be performed by using Jacobian in cases where Chebychev–Grübler–Kutzbach criterion cannot [@GOGU2005427] accurately determine the mobility. As a result, Jacobian is a significant part for both kinematic analysis, dimensional synthesis and mobility determination of a mechanism. Hence, the formulation of Jacobian has its key importance in the literature and in the application of performance optimisation along with mobility computation. Formulation of Jacobian for serial manipulators can be computed easily, however, it is increasingly complicated to formulate Jacobian for parallel manipulators due to the existence of passive joint velocities and the nature in which these are related to active joint velocities. Several studies [@606737;@846382;@1220722;@ALTUZARRA20061407] exist for formulation of parallel manipulators but all these studies have dependency on human inspection at some level. Several open source softwares are also available for Jacobian formulation [@nadeau2019pybotics;@Owan2018;@Baumgärtner2022;@Lee2018], but either their application is limited to serial manipulators or they require human intervention or they are part of computationally expensive simulations. For example, TriP [@Baumgärtner2022] does facilitate Jacobian computation with closed-loop chains in the manipulator structure, however, as obvious from the triped robot example in the documentary, human inspection is apparently required to appropriately join the legs of the robot. To alleviate the drawback of the requirement of human inspection, the present research aims to formulate Jacobian that is required for dimensional synthesis for optimal performance around a single point that can be used for any non-redundant manipulator without any dependency on human inspection. Jacob and Dasgupta [@synthesis] used a systematic method as a tool to formulate Jacobian matrices for several manipulators in bulk for performance optimisation around a given task point. However, several steps in that algorithm are not totally computerised but rather human-intervention-dependent. Moreover, it can be applicable only with four types of joints. The current paper extends their method to present a fully computerisable Jacobian formulation algorithm that is applicable for general non-redundant planar and spatial manipulators of any topological structure of seven types of joints, namely revolute, prismatic, cylindrical, spherical, universal, helical and plane joints. Based on this extended method, the Python package `Automatic Computation for Robot Design (ACRoD)` is developed by the authors. `ACRoD` provides a Python-based package for generating functions required to compute Jacobian at a given configuration for a given end-effector point, merely from the simple topological information of the robot in a fully automated manner. This can be directly used in optimisation process to derive optimal dimensions of the robot for optimal performance around a given end-effector point, thereby avoiding many tedious steps in manual formulation, especially when a comparison study is performed on multiple manipulators in bulk [@synthesis].
[…]
