### 2023-12-04 046c7bf
Update 3R_serial_robot.md
