+++ Robot_Topology_Matrix.md
@@ -21,6 +21,22 @@ In a serial manipulator, typically all joints are active. However, in parallel m
 
 In ACRoD Jacobian formulation, only revolute and prismatic joints are considered for actuation and hence the other types joints (cylindrical, spherical, universal, helical and planar) can only be passive.
 
+## Examples
+### 3R Planar Serial Robot
+Let us consider the 3R planar serial robot as shown in (a) of the figure below.
+
+<p align="center">
+    <img src="../images/RRR.png" alt="RRR planar serial manipulator" width="500px">
+</p>
+
+The above robot has four links (including the fixed link). Hence the Robot-Topology matrix $M$ would be a $4\times4$ matrix, with all of its diagonal elements as '9's. Since the base link is considered to be the first link and the end-effector link is considered to be the last link, the base and the end-effector links are assigned the names $L_1$ and $L_4$, as shown in (b) of the above figure. Now that the base and the end-effector links are fixed, the other two links can be arbitrarily named. Since there are two links left out, we have $2!=2$ possible ways of naming them. One way is to name the link that is connected to the base link as $L_2$ and the link that is connected to the end-effector link as $L_3$, as shown in (c) of the above figure. Here, we have three revolute joints between $L_1$ \& $L_2$, $L_2$ \& $L_3$ and $L_3$ \& $L_4$. Hence, the pairs of row-column indices of the Robot-Topology matrix in the ascending order, i.e., (1,2), (2,3) and (3,4), would have the corresponding elements of the Robot-Topology matrix as '1's. From (a) of the figure, it can be observed by the specified arrows that all the three revolute joints are actuated (as expected in a non-redundant serial robot). Hence, the elements corresponding to the indices of these three actuators (joints) in descending order, i.e., (2,1), (3,2) and (4,3), would be '1's. Since there is no joint other than these three joints, the rest of the elements of the Robot-Topology matrix would be zeros. Hence, the Robot-Topology matrix would be as shown below.
+
+$$\bf{M} = \left[\begin{matrix}9 & 1 & O & O \\\\1 & 9 & 1 & O\\\\O & 1 & 9 & 1\\\\O & O & 1 & 9\end{matrix}\right]$$
+
+Another way is to name the link that is connected to the base link as $L_3$ and the link that is connected to the end-effector link as $L_2$, as shown in (d) of the above figure. In that case, the elements (1,3), (2,3) and (2,4) would be '1's, indicating the existence of revolute joints. The elements (3,1), (3,2) and (4,2) would be '1's, indicating that the corresponding joints, i.e., (1,3), (2,3) and (2,4), are actuators. Hence, the Robot-Topology matrix would be as shown below.
+
+$$\bf{M} = \left[\begin{matrix}9 & 0 & 1 & O \\\\0 & 9 & 1 & 1\\\\1 & 1 & 9 & 1\\\\O & O & 0 & 9\end{matrix}\right]$$
+
 ## References
 <a id="1">[1]</a> 
 Jacob, Akkarapakam Suneesh, Bhaskar Dasgupta, and Rituparna Datta. "Enumeration of spatial manipulators by using the concept of Adjacency Matrix." arXiv preprint arXiv:2210.03327 (2022).
