+++ misc/Robot_Topology_Matrix.md
@@ -48,13 +48,22 @@ Let us consider the 4R-4P planar serial-parallel hybrid robot as shown in (a) of
 
 The above robot has seven links. Hence the Robot-Topology matrix $M$ would be a $7\times7$ matrix, with all of its diagonal elements as '9's. As per the convention, the base and the end-effector links are assigned the names $L_1$ and $L_7$ respectively. One possible way of naming the other links is shown in (b) of the above figure. Here, the robot has four prismatic joints between $L_1$ \& $L_2$, $L_1$ \& $L_3$, $L_2$ \& $L_4$ and $L_3$ \& $L_4$. Hence, the pairs of row-column indices of the Robot-Topology matrix in the ascending order, i.e., (1,2), (1,3), (2,4) and (3,4), would have the corresponding elements of the Robot-Topology matrix as '2's. Likewise, the robot has four revolute joints between $L_4$ \& $L_5$, $L_4$ \& $L_6$, $L_5$ \& $L_7$ and $L_6$ \& $L_7$, and hence, the elements at (4,5), (4,6), (5,7) and (6,7) would be '1's. From (a) of the figure, it can be observed by the specified arrows that the joints between $L_1$ \& $L_2$, $L_1$ \& $L_3$ and $L_4$ \& $L_5$ are actuators. Hence, the elements corresponding to the indices of these three actuators (joints) in descending order, i.e., (2,1), (3,1) and (5,4), would be '1's. The rest of the elements of the Robot-Topology matrix would be zeros. Hence, the Robot-Topology matrix would be as shown below.
 
-$$\bf{M} = \left[\begin{matrix}9 & 1 & 0 & 0 \\\\1 & 9 & 1 & 0\\\\0 & 1 & 9 & 1\\\\0 & 0 & 1 & 9\end{matrix}\right]$$
+$$\bf{M} = \left[\begin{matrix}9 & 2 & 2 & 0 & 0 & 0 & 0\\\\1 & 9 & 0 & 2 & 0 & 0 & 0\\\\1 & 0 & 9 & 2 & 0 & 0 & 0\\\\0 & 0 & 0 & 9 & 1 & 1 & 0\\\\0 & 0 & 0 & 1 & 9 & 0 & 1\\\\0 & 0 & 0 & 0 & 0 & 9 & 1\\\\0 & 0 & 0 & 0 & 0 & 0 & 9\end{matrix}\right]$$
 
-Another way is to name the link that is connected to the base link as $L_3$ and the link that is connected to the end-effector link as $L_2$, as shown in (d) of the above figure. In that case, the elements (1,3), (2,3) and (2,4) would be '1's, indicating the existence of revolute joints. The elements (3,1), (3,2) and (4,2) would be '1's, indicating that the corresponding joints, i.e., (1,3), (2,3) and (2,4), are actuators. Hence, the Robot-Topology matrix would be as shown below.
+Other possible Robot-Topology matrices can be framed (by changing the numbering of links), which would be equivalent to the above matrix.
 
-$$\bf{M} = \left[\begin{matrix}9 & 0 & 1 & 0 \\\\0 & 9 & 1 & 1\\\\1 & 1 & 9 & 0\\\\0 & 1 & 0 & 9\end{matrix}\right]$$
+### RSSR-SSR Spatial Parallel Robot
+Let us consider the RSSR-SSR spatial parallel robot as shown in (a) of the figure below.
 
-Either matrix can be used as Robot-Topology matrix.
+<p align="center">
+    <img src="./RSSRSSR.png" alt="RSSR-SSR spatial parallel robot" width="800px">
+</p>
+
[…]
