+++ examples/Jacobian/codes/RRR_code.md
@@ -14,7 +14,7 @@ $$\bf{M} = \left[\begin{matrix}L_1 & R & O & O \\\\A & L_2 & R & O\\\\O & A & L_
 
 ### Jacobian for planar manipulators
 
-The topological information of a robot is to be specified by using its robot-topology matrix, as defined [here](../../../Robot_Topology_Matrix.md). For RRR planar serial manipulator shown above, the robot topology matrix is given by
+The topological information of a robot is to be specified by using its robot-topology matrix, as defined [here](./Robot_Topology_Matrix.md). For RRR planar serial manipulator shown above, the robot topology matrix is given by
 
 $$\left[\begin{matrix}9 & 1 & 0 & 0 \\\\ 1 & 9 & 1 & 0 \\\\ 0 & 1 & 9 & 1 \\\\ 0 & 0 & 1 & 9\end{matrix}\right]$$
 
@@ -55,7 +55,7 @@ In an ipynb file of JupyterLab, the above code would produce the following outpu
 
 $$\left[\begin{matrix}- a_{y} + r_{(1,2)y} & - a_{y} + r_{(2,3)y} & - a_{y} + r_{(3,4)y}\\\\a_{x} - r_{(1,2)x} & a_{x} - r_{(2,3)x} & a_{x} - r_{(3,4)x}\\\\1 & 1 & 1\end{matrix}\right]$$
 
-The above Jacobian is based on the notations defined and described [here](../../../Notation_and_Nomenclature.md).
+The above Jacobian is based on the notations defined and described [here](../../../Notation_and_Nomenclature).
 
 Active joint velocities, in the corresponding order, can be viewed by running the following lines.
 ```py
