+++ paper/paper.md
@@ -31,9 +31,9 @@ The Jacobian of a robot refers to the matrix that linearly maps the velocity com
 
 # Statement of need
 
-For a manipulator of a given topology, designing the dimensions based on optimising Jacobian-based performance parameters (such as manipulability index and condition number [@patel2015manipulator]) around a given end-effector point would require only the topological information for the formulation of Jacobian, as every other step can be automated. Formulation of Jacobian for parallel manipulators and serial-parallel hybrid manipulators are non-trivial, although all the steps of Jacobian formulation even in those cases would have to stem from the mere information of topology of the robot. ACRoD automates the non-trivial formulation of Jacobian systematically. It uses a matrix-based representation of the topology of the robotic manipulator (referred to here as the robot-topology matrix, of which more information is provided [here](https://github.com/suneeshjacob/ACRoD/blob/main/Robot_Topology_Matrix.md)) which is a modified version of the graph adjacency matrix representation [@enumeration] of robotic manipulators. This Jacobian formulation can be used to generate numerical Jacobian matrices with a few random configurations, from which the singular values can be calculated which can confirm the Degree of Freedom (DoF). In other words, the DoF of a given robot topology for a given base link and a given end-effector link can be verified by using this Jacobian function even in cases where Chebychev–Grübler–Kutzbach criterion fails to verify. This can be useful in mechanism synthesis to accurately verify the mobility of a given manipulator directly from its robot-topology matrix by using the method shown in Yang et al.'s paper [@YANG20081175].
+For a manipulator of a given topology, designing the dimensions based on optimising Jacobian-based performance parameters (such as manipulability index and condition number [@patel2015manipulator]) around a given end-effector point would require only the topological information for the formulation of Jacobian, as every other step can be automated. Formulation of Jacobian for parallel manipulators and serial-parallel hybrid manipulators are non-trivial, although all the steps of Jacobian formulation even in those cases would have to stem from the mere information of topology of the robot. ACRoD automates the non-trivial formulation of Jacobian systematically. It uses a matrix-based representation of the topology of the robotic manipulator (referred to here as the robot-topology matrix, of which more information is provided [here](https://github.com/suneeshjacob/ACRoD/blob/main/misc/Robot_Topology_Matrix.md)) which is a modified version of the graph adjacency matrix representation [@enumeration] of robotic manipulators. This Jacobian formulation can be used to generate numerical Jacobian matrices with a few random configurations, from which the singular values can be calculated which can confirm the Degree of Freedom (DoF). In other words, the DoF of a given robot topology for a given base link and a given end-effector link can be verified by using this Jacobian function even in cases where Chebychev–Grübler–Kutzbach criterion fails to verify. This can be useful in mechanism synthesis to accurately verify the mobility of a given manipulator directly from its robot-topology matrix by using the method shown in Yang et al.'s paper [@YANG20081175].
 
[…]
