### 2023-12-04 suneeshjacob/ACRoD#9 @suneeshjacob
Hi, I made appropriate changes. Regarding adding technical information to the paper, there is a lot of technical information and I think that would occupy a lot of space in the paper. I first planned to put the full-length paper (with all the mathematical jargons) into `paper.md` but after i read [the docs](https://joss.readthedocs.io/en/latest/submitting.html) where it says full length paper is not permitted, i gave up the idea and made this short paper.

Summary of the changes:
1) A subsection is added in the paper for comparison with other software.
2) Moved the second paragraph under a new subsection named 'Method'.
3) Moved many files to a new folder named 'misc', so that the links of the files can stay more permanent.
4) Fixed the nitpicks (thanks)!
