### 2023-12-04 #5927 @suneeshjacob
@JHartzer i made appropriate changes based on your comments.

Here is a summary of changes made:
1) A subsection is added in the paper for comparison with other software.
2) Moved the second paragraph under a new subsection named 'Method'.
3) Moved many files to a new folder named 'misc', so that the links of the files can stay more permanent.
4) Fixed the nitpicks (thanks)!
