+++ README.md
@@ -33,13 +33,13 @@ The corresponding Jacobian function can be formulated as follows.
 Firstly, the required functions are imported as shown below.
 ```py
 from acrod.functions import jacobian
-from numpy import matrix
+from numpy import array
 ```
 
 
 The robot-topology matrix for 3R planar serial manipulator is defined and jacobian information is processed via the imported jacobian class as follows.
 ```py
-M = numpy.array(
+M = array(
         [[9, 1, 0],
          [1, 9, 1],
          [0, 1, 9]]
