+++ README.md
@@ -143,7 +143,7 @@ initial_guess = [1,1]
 res = minimize(condition_number, initial_guess)
 ```
 
-The link lengths $l_2$ and $l_3$ are given by $l_2 = \left|\textbf{r}\_{12}-\textbf{r}\_{23}\right|$ and $l_3 = \left|\textbf{r}\_{23}-\textbf{a}\right|$. By using the code below, the link lengths of 2R robot can be computed.
+The link lengths $l_2$ and $l_3$ are given by $l_2 = \lVert \textbf{r}\_{12}-\textbf{r}\_{23} \rVert$ and $l_3 = \lVert\textbf{r}\_{23}-\textbf{a}\rVert$. By using the code below, the link lengths of 2R robot can be computed.
 
 ```py
 l1 = np.linalg.norm(res.x)
