+++ README.md
@@ -147,7 +147,7 @@ from numpy import hstack, random
 end_effector_point = [1,2]
 base_reference_point = [0,0]
 r12 = base_reference_point
-jac_fun = lambda y: jacobian_function(end_effector_point, hstack((end_effector_point,initial_guess))
+jac_fun = lambda y: jacobian_function(end_effector_point, hstack((end_effector_point,initial_guess)))
 condition_number = lambda z: condition_number_func(jac_fun(z))
 initial_guess = random.rand(len(jac.parameters)-len(base_reference_point))
 res = minimize(condition_number, initial_guess)
