### 2023-12-09 dfdbf9e
Added checks on logical consistency of Robot-Topology matrix
