### 2023-12-09 #5927 @suneeshjacob
> ## Summary
> The main stated contribution of this library is to provide a framework for the automatic calculation of the Jacobian in the case of several robot topologies. The aim is to simplify and automatize the tedious computations to obtain the jacobian of the end-effector. The authors present the procedure in `algorithm1.png` for this computation, based on an augmented adjacency matrix defining the robot topology.
>
> ## Main comments
> Several points need to be addressed in order to increase the quality of the contribution:
>
> * How does your library place with respect to current solutions?
>
>   1. Several libraries implement algorithms to assess performance metrics of the end-effector (jacobian is just an intermediate step). For instance `PyCapacity`, another JOSS contribution, aims at solving a similar problem by the use of polytopes and ellipsoids. To this end, I was expecting the automatic computation of at least a few manipulability indicators (e.g. eigenvalues, matrix norms, condition number, et cetera).
>   2. Several robotic libraries (e.g. `pinocchio`) compute the jacobian of the end-effector and can be easily used to get closed-loop jacobians. In the case of closed-loop kinematic chains, if the system is not degenerate, by the use of the implicit function theorem, it is possible to express the passive velocities as a function of the independent ones. This approach is detailed, for instance in [this paper](https://hal.science/hal-04191553/document), which also introduces useful performance metrics.
> * The statement that the design of a robot can be automatized by the use of this library needs to be proven with examples.  Also the title of the paper "Automatic computation for robot design" has little to do with the contribution scope, and this may be confusing for the potential users of the tool. At the current stage, it seems that no form of automatic design generation or optimization (this can be quickly checked as no optimization library is required).
> * The methodology heavily uses a custom encoding of the topology explained in the readme `Robot_Topology_Matrix`. This mathematical object is misleadingly called an adjacency matrix (as, following the reasoning, it may even be nonsymmetric). In reality, the described matrix is a custom encoding of the robot topology and joint types (where upper and lower diagonal elements encode the joint type and if it is associated with an independent variable). For more complex robots it may become very error-prone, and it seems to me that no checks on the matrix (see following comments) are made before manipulating it. Would the library benefit from the use of more standardized formats like `urdf` or `srdf` instead?
>
> ## Code/implementation details
> On the library: - the code generally has comments, but the repo is missing examples. - more modularity in the code structure would be appreciated, for now, there is just a generic `functions.py` which provides all the repo functionality. - notes on `Robot_Topology_Matrix.md`: - Repetition of "The adjacency matrix is normally a symmetric matrix" - What happens if the system has multiple end effectors? Do you need to permutate the matrix at each time? - How do you treat the constraints arising from between active passive joints? To be more clear, once you have a kinematic loop, not all joints can be active, and the difference between active and passive joints needs to equal the free DoF of the chain. - When explaining the meaning of the lower diagonal part of the matrix I see a contradiction. On one side the goal of the library is to allow the automatic computation of the jacobian, but as pointed out there is some freedom in the specification of the lower part. This is indeed a cumbersome step that needs additional insights.
>
> ## Testing
> Unfortunately, I could not make use of the library in some simple cases. The example runs but testing the library for a serial revolute joint manipulator does not give the expected outputs. Moreover, it seems that not all the joint types are functional. See the driving code [here](https://gist.github.com/gfadini/6a99978064fc51b24d0327e84fcc4dec)

@gfadini hi, here is my response to your comments of the paper:

Main comments:

1. (i) The objective of the current implementation is to formulate the Jacobian, which itself has a lot of tedious processes that were automated. PyCapacity is a package for extracting performance parameters from an already existing Jacobian, and hence not a competitor to ACRoD. The reason why performance parameters are not included in this repository is that performance parameters are often chosen by the designers based on multiple factors. For example, `PyCapacity` provides polytope-based measure of manipulability which is more computationally expensive to compute but has more accurate quantification of performance, whereas ellipsoid-based measure of manipulability can be computed by simply computing $\det{\left(J^TJ\right)}$ but describes less accurate quantification of performance when compared with polytope-based measure. If the designer needs quick results then it would be more appropriate to go for the ellipsoidal approach and if more accurate measure is needed then the polytope approach would be more appropriate. Also, performance parameters such as condition number may suffer from dimensional inhomogenity due to the mix of both linear and angular velocities that have different units. There are several methods discussed in the literature to tackle this issue. One method is to consider a characteristic length to construct a scaling matrix, but computation of this characteristic length has arbitrariness. Separating the linear and angular blocks of Jacobian matrix and computing the product of condition numbers of individual blocks is another approach that can avoid the issue of arbitrariness but it could be computationally more expensive. So, it ultimately depends on the choice and specific requirements of the designer.
(ii) The library [pinocchio](https://github.com/stack-of-tasks/pinocchio/tree/master) that you mentioned, which is a C++ based library, does not mention anywhere explicitly about its inclusion of Jacobian for closed-loop kinematic chains in its documentation (even though it may be able to handle closed-loop chains). Since its base language is C++, I couldn't assess much about how it handles or if it ever handles in the first place. However, the usage of implicit function theorem that you mentioned, suggests that it is based on numerical differentiation of nonlinear positions (implicit position equations) of the robot (as mentioned by \Phi(x) in the paper that you tagged). If that's the case then that would be unnecessarily invoking position equations, as mere Jacobian does not require listing of nonlinear implicit position equations. Moreover, `ACRoD` can handle degenerate equations of non-redundant manipulators in two steps, by framing the connectivity matrices for linear and angular velocities followed by checking the rank of the matrices and finding the corresponding pivot elements to reduce the equations to full-rank equations. The 4R-4P Serial-Parallel Hybrid Robot as an example is chosen to highlight this feature that it automatically finds and reduces the system into full-rank equations to avoid degeneracy. And even if all of these were in `pinocchio` library, it would still be the case that `ACRoD` could automatically generate the Jacobian functions simply from the Robot-Topology matrix which is apparently much simpler way of describing the topological information of the robot than modelling each link and the connections of links in `pinocchio`. By the way, so far in the documentation of `pinocchio` I could only see loading robots from URDF (which is limited to serial manipulators) or SRDF and I am yet to see any method to build customised robots in `pinocchio`. And if it were the case that `pinocchio` relies on URDF and SRDF then it might be required to produce unnecessarily big files (such as URDF and SRDF) to specify the pose of a robot each time while doing optimisation, as in dimensional synthesis the link lengths would be variables.
While this is the case with `pinocchio`, the paper you tagged apparently used the author's own package [closed_loop_kinematics_pinocchio](https://gitlab.laas.fr/vbatto/closed_loop_kinematics_pinocchio), in which the author used `pinocchio` package but developed his/her own script for actually handling closed-loop kinematic linkages, apparently. Even `closed_loop_kinematics_pinocchio` doesn't seem to have proper documentation and moreover it looks like the code is not properly written (for example, in `closed_loop_kinematics_pinocchio/utils/closed_loop_jacobian.py` there is a function named `jacobianFinitDiffClosedLoop` which uses this function named `closedLoopForwardKinematics` but its definition is not found either in the file or in `closed_loop_kinematics.py`), hence it is hard to understand the logic of the code. However, from the structure of the code, it seems to be computing Jacobian by using Finite Difference Method. It is possible to compute Jacobian by differentiating nonlinear (implicit) position equations by time and rewriting passive joint velocities (derivatives of displacements of passive joints with respect to time) in terms of active joint velocities (derivatives of displacements of active joints with respect to time). However, it is not clear how the degeneracy of equations is handled. And even if it is handled properly, it still uses `pinocchio` model to compute Jacobian, which means the complexity in specifying a robot's topology would apparently arise here as well. Using a huge file (such as URDF or SRDF) for merely specifying the topological information would be unnecessary, whereas a simple Robot-Topology matrix would make it much simpler.
To summarise, the documentation of `pinocchio` apparently neither explicitly mentions nor claims the inclusion of closed-loop linkages in computing Jacobian. However, the package `closed_loop_kinematics_pinocchio` seems to be able to include closed-loop linkages. However, these have more complexity in specifying robot's topology (as input), which is much simpler in ACRoD. Thus, to the best of my knowledge, ACRoD stands above these two packages in the context of formulating Jacobian for dimensional synthesis from mere topology, although other packages may stand above ACRoD in other contexts.

2. Yes, it is true that currently the repository does not use any optimization library. However, the aim of the repository is to provide computational equipment that would be useful in the process of dimensional synthesis aimed at ultimately designing robots. Nevertheless, I now added the script for dimensional synthesis in `readme.MD` for the 2R Planar Serial Robot example, where the optimal link lengths (under some assumptions) are ultimately found for operation around a given end-effector point. While the repository's current title for the repository is chosen based on the long-term plan that the authors have with the repository, the authors are open to changing the title of the paper/repository if the change can help reduce any confusion for potential users.

3. Yes it is true that the input matrix is not exactly an adjacency matrix, that is precisely why it is called Robot-Topology matrix rather than adjacency matrix. It is a modified version of adjacency matrix, which has all the information that an adjacency matrix would hold, and additionally it holds the information about what joints are actuators and what joints are passive. Yes, earlier there were no checks on the Robot-Topology matrix, however I now added a function that automatically checks for logical consistency of Robot-Topology matrix before initialising Jacobian class of the robot, which throws appropriate errors if the matrix is found to be inconsistent. URDF or SRDF wouldn't be appropriate to replace the Robot-Topology matrix here because firstly URDF apparently does not have support for closed-loop linkages [1], and secondly even if SRDF does (which i am not sure of), it would be involving specification of positions of joints, orientations of joints (and perhaps collision information and inertia information) which are not necessary and would only be arbitrary while occupying space.

[1] Tola, D., & Corke, P. (2023). Understanding URDF: A survey based on user experience. 2023 IEEE 19th International Conference on Automation Science and Engineering (CASE), 1–7. https://doi.org/10.1109/CASE56687.2023.10260660

Code/implementation details

On the library:
1) Examples are added.
2) The currently available functions are in `functions.py`, but this file encloses all the functions required for this particular feature of Jacobian formulation around a single end-effector point. The authors have a future plan to provide multiple sub-modules of ACRoD, including facilitation of multiple end-effector points, positional kinematics, dynamics, link flexibility, etc., which would be having the modular structure.
3. `Robot_Topology_Matrix.md` is modified with examples.
4. Removed the repetition (thanks)!
5. The case of multiple end-effectors is not within the scope of this implementtion. However, a particular link in the matrix can be chosen as the end-effector link by bringing that link to the last diagonal element of the matrix and adjusting the off-diagonal elements to preserve the topology. If all possible end-effector links for the manipulator are to be enumerated then in such a case permutation of diagonal elements in the matrix would be needed.
6. The constraints arising from between active and passive joints are handled by framing $A\theta=0$ and splitting the system of equations into two parts corresponding to active and passive joint velocities, and finally expressing the passive joint velocities in terms of active joint velocities. This step is explained [here](https://github.com/suneeshjacob/ACRoD/blob/main/misc/Mathematics_behind_Jacobian_formulation.md) in the repository.
7. I don't see a contradiction there. The lower diagonal part of the matrix is dedicated for specifying whether the joints are actuators or not, because the dimensional synthesis could depend on which joints are actuated in a robot. It is possible for the optimal performance to change when different joints are actuated. For example, for a 3-RPR planar parallel manipulator, the actuators could be the three prismatic joints but it is also possible for the actuators to be three revolute joints. And in each case the dimensional synthesis could yield different optimal dimensions of the robot, as the Jacobian structure changes in each case. Hence, it would be in the interest of the designer to provide which joints to be actuated. The goal of the library to automate the computation of the Jacobian from the specified deterministic robot topology including labelling actuators. The lack of labelling actuators would lead to ambiguity of which joints to be actuated, thereby causing ambiguity in formulating the Jacobian itself.

Testing
1. You are right, it didn't work with 2-P planar robot. There was a bug in the code regarding the angular velocities of paths with prismatic joints alone (if a path has prismatic joints alone then the angular velocity through that path would be zero, which was omitted as trivial, hence accessing angular velocity equations threw a `IndexError: list index out of range` error). This bug is fixed. Now it is working, [here](https://gist.github.com/suneeshjacob/f8436d79ecd64ad055a81e3193f18859) is the code implementation. Thanks for pointing it out, by the way!
2. The reason why it failed for joint types 3,4,5,6 and 7, however, is not due to a bug in the code. It is due to the restrictive consideration in this implementation that only revolute and prismatic joints would serve as actuators. The other five types of joints (i.e., cylindrical, spherical, universal, helical and plane joints) can only occur as passive joints. This makes sense because in practical robots because due to the availability of electric motors and pneumatic actuators facilitating revolute and prismatic joints respectively, the other five types of joints are available almost never used as passive joints in a meaningful sense that falls outside the scope of combinations of revolute and prismatic joints. In your `augmentedAdjacency` example, all the cases are referring to serial manipulators. In non-redundant serial manipulators, all joints need to be active joints and hence due to the inability of joint types 3,4,5,6 and 7 to be actuating joints, the build is failing. If you use an example matrix in which these joint types are used as passive joints, it should work without any errors.
