### 2023-12-27 suneeshjacob/ACRoD#10 @CameronDevine
Convert Examples to Notebooks
The example are looking much better. The main problem now, is that they cannot be run directly. The format currently looks very similar to a notebook, but is not in a notebook file format. I would suggest moving the examples to Jupyter notebooks to make them easier to run, edit, and experiment with in general.
