### 2023-12-27 suneeshjacob/ACRoD#11 @CameronDevine
Package Documentation
There does not seem to be a centralized location to find the documentation of all functions. The usage of many of the functions is provided in various examples, but an easily accessible and centralized place to find the full documentation of the package is unavailable. As all functions appear to have docstrings, it should be simple to use a tool such as Sphinx to create an accessible source of information about the package. A service such as [readthedocs.io](http://readthedocs.io) could also be used.
