### 2023-12-28 195891d
Update RSSRSSR_code.md
Fixed a mistake
