### 2023-12-29 ac05ac4
Renamed functions folder to 'jacobian', and renamed jacobian class to 'Jacobian'
Renamed functions folder to 'jacobian', and renamed jacobian class to 'Jacobian', and made subsequent changes in the examples, etc.
