+++ examples/Jacobian/codes/4R4P_code.md
@@ -22,7 +22,7 @@ The corresponding Jacobian function can be formulated as follows.
 
 Firstly, the required functions are imported as shown below.
 ```py
-from acrod.functions import jacobian
+from acrod.jacobian import Jacobian
 from numpy import array
 ```
 
@@ -38,7 +38,7 @@ M = array(
          [0, 0, 0, 0, 0, 9, 1],
          [0, 0, 0, 0, 0, 0, 9]]
     )
-jac = jacobian(M, robot_type = 'planar')
+jac = Jacobian(M, robot_type = 'planar')
 ```
 
 
