+++ examples/Jacobian/codes/RSSRSSR_code.md
@@ -22,7 +22,7 @@ The corresponding Jacobian function can be formulated as follows.
 
 Firstly, the required functions are imported as shown below.
 ```py
-from acrod.functions import jacobian
+from acrod.jacobian import Jacobian
 from numpy import array
 ```
 
@@ -37,7 +37,7 @@ M = array(
          [0, 0, 0, 0, 9, 4],
          [0, 0, 0, 0, 0, 9]]
     )
-jac = jacobian(M)
+jac = Jacobian(M)
 ```
 
 
