+++ docs/acrod.rst
@@ -1,0 +1,21 @@
+acrod package
+=============
+
+Submodules
+----------
+
+acrod.jacobian module
+---------------------
+
+.. automodule:: acrod.jacobian
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: acrod
+   :members:
+   :undoc-members:
+   :show-inheritance:
