+++ docs/modules.rst
@@ -1,0 +1,7 @@
+acrod
+=====
+
+.. toctree::
+   :maxdepth: 4
+
+   acrod
