+++ docs/index.rst
@@ -1,5 +1,5 @@
 .. ACRoD documentation master file, created by
-   sphinx-quickstart on Sat Dec 30 16:23:04 2023.
+   sphinx-quickstart on Sat Dec 30 17:00:29 2023.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
