+++ requirements.txt
@@ -1,2 +1,5 @@
 sympy
 numpy
+sphinx==5.3.0
+sphinx_rtd_theme==1.1.1
+readthedocs-sphinx-search==0.1.1
