+++ docs/conf.py
@@ -22,7 +22,7 @@ copyright = '2023, Akkarapakam Suneesh Jacob and Rituparna Datta'
 author = 'Akkarapakam Suneesh Jacob and Rituparna Datta'
 
 # The full version, including alpha/beta/rc tags
-release = '1.1.12'
+release = '1.1.16'
 
 
 # -- General configuration ---------------------------------------------------
@@ -51,7 +51,7 @@ exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'alabaster'
+html_theme = 'sphinx_rtd_theme'
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
