+++ requirements.txt
@@ -1,5 +1,2 @@
 sympy
 numpy
-sphinx==5.3.0
-sphinx_rtd_theme==1.1.1
-readthedocs-sphinx-search==0.1.1
