+++ docs/requirements.txt
@@ -1,3 +1,4 @@
 sphinx==5.3.0
 sphinx_rtd_theme==1.1.1
 readthedocs-sphinx-search==0.1.1
+acrod
