+++ docs/installation_and_usage.rst
@@ -1,3 +1,6 @@
+Installation and Usage
+======================
+
 Installation
 ------------
 
