### 2023-12-30 a41a739
Update Notation_and_Nomenclature.md
