### 2023-12-30 1721025
Update Notation_and_Nomenclature.md
