+++ docs/modules.rst
@@ -1,5 +1,5 @@
-acrod
-=====
+API Documentation
+=================
 
 .. toctree::
    :maxdepth: 4
