+++ docs/mathematics_behind_jacobian_formulation.rst
@@ -1,5 +1,8 @@
+Mathematics Behind Jacobian Formulation
+=======================================
+
 Steps to formulate Jacobian
-===========================
+---------------------------
 
 This method is an extension/modification of the method explained
 presented in `[1] <#1>`__. The approach that is typically used to
@@ -245,29 +248,49 @@ serial manipulators is as shown in equation (8).
 The above steps of the algorithm are concisely shown in the pseudocode
 of algorithm 1.
 
-.. raw:: html
+.. image:: ../misc/algorithm1.png
+   :alt: Alternative Text
+   :width: 300
+   :align: center
 
-   <p align="center">
+.. image:: ../misc/algorithm2.png
+   :alt: Alternative Text
+   :width: 300
+   :align: center
 
-.. raw:: html
+.. image:: ../misc/algorithm3.png
+   :alt: Alternative Text
+   :width: 300
+   :align: center
 
-   </p>
+.. pcode::
+   :linenos:
 
-.. raw:: html
-
-   <p align="center">
-
-.. raw:: html
-
-   </p>
-
-.. raw:: html
-
-   <p align="center">
-
-.. raw:: html
-
-   </p>
+    % This quicksort algorithm is extracted from Chapter 7, Introduction to Algorithms (3rd edition)
+    \begin{algorithm}
+    \caption{Quicksort}
+    \begin{algorithmic}
+    \PROCEDURE{Quicksort}{$A, p, r$}
+        \IF{$p < r$}
+            \STATE $q = $ \CALL{Partition}{$A, p, r$}
+            \STATE \CALL{Quicksort}{$A, p, q - 1$}
+            \STATE \CALL{Quicksort}{$A, q + 1, r$}
+        \ENDIF
+    \ENDPROCEDURE
+    \PROCEDURE{Partition}{$A, p, r$}
+        \STATE $x = A[r]$
+        \STATE $i = p - 1$
+        \FOR{$j = p$ \TO $r - 1$}
+            \IF{$A[j] < x$}
+                \STATE $i = i + 1$
+                \STATE exchange
+                $A[i]$ with     $A[j]$
+            \ENDIF
+            \STATE exchange $A[i]$ with $A[r]$
+        \ENDFOR
+    \ENDPROCEDURE
+    \end{algorithmic}
+    \end{algorithm}
 
 This algorithm can be used to find the four matrices
 :math:`\mathbf{J_a}`, :math:`\mathbf{J_p}`, :math:`\mathbf{A_a}` and
@@ -278,7 +301,7 @@ the Jacobian would simply be
 :math:`\mathbf{\widetilde{J}}=\mathbf{J_a}`.
 
 References
-==========
+----------
 
 [1] Jacob, Akkarapakam Suneesh, and Bhaskar Dasgupta. “Dimensional
 synthesis of spatial manipulators for velocity and force transmission
