+++ docs/requirements.txt
@@ -2,4 +2,3 @@ sphinx==5.3.0
 sphinx_rtd_theme==1.1.1
 readthedocs-sphinx-search==0.1.1
 acrod
-sphinxcontrib-pseudocode
