+++ docs/RSSRSSR_code.rst
@@ -1,0 +1,407 @@
+RSSR-SSR Spatial Parallel Robot
+===============================
+
+Figure
+------
+
+.. image:: ../examples/Jacobian/images/RSSRSSR.png
+   :alt: Alternative Text
+   :width: 300
+   :align: center
+
+The RSSR-SSR Serial-Parallel Hybrid Robot is taken from `[1] <#1>`__,
+and a picture of the robot is shown in the above figure. The
+corresponding adjacency matrix is given by
+
+.. math:: \bf{M} = \left[\begin{matrix}L_1 & R & R & O & R & O\\A & L_2 & O & O & O & S\\A & O & L_3 & S & O & O\\O & O & O & L_4 & O & S\\O & O & O & O & L_5 & S\\O & O & O & O & O & L_6\end{matrix}\right]
+
+Usage
+-----
+
+Jacobian for planar manipulators
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+The topological information of a robot is to be specified by using its
+robot-topology matrix, as defined
+`here <../../../misc/Robot_Topology_Matrix.md>`__. For RSSR-SSR spatial
+parallel manipulator shown above, the robot topology matrix is given by
+
+.. math:: \left[\begin{matrix}9 & 1 & 1 & 0 & 1 & 0\\1 & 9 & 0 & 0 & 0 & 4\\1 & 0 & 9 & 4 & 0 & 0\\0 & 0 & 0 & 9 & 0 & 4\\0 & 0 & 0 & 0 & 9 & 4\\0 & 0 & 0 & 0 & 0 & 9\end{matrix}\right]
+
+The corresponding Jacobian function can be formulated as follows.
+
+Firstly, the required functions are imported as shown below.
+
+.. code:: py
+
+   from acrod.jacobian import Jacobian
+   from numpy import array
+
+The robot-topology matrix for RSSR-SSR spatial parallel manipulator is
+defined and jacobian information is processed via the imported jacobian
+class as follows.
+
+.. code:: py
+
+   M = array(
+           [[9, 1, 1, 0, 1, 0],
+            [1, 9, 0, 0, 0, 4],
+            [1, 0, 9, 4, 0, 0],
+            [0, 0, 0, 9, 0, 4],
+            [0, 0, 0, 0, 9, 4],
+            [0, 0, 0, 0, 0, 9]]
+       )
+   jac = Jacobian(M)
+
+Jacobian function is generated as shown below.
+
+.. code:: py
+
+   jacobian_function = jac.get_jacobian_function()
+
+In the process of generating the above jacobian function, other
+attributes of the jacobian object also are updated. Symbolic Jacobian
+matrices can be extracted from the attributes. Since this is a
+non-serial robot, there would be four matrices required to compute the
+Jacobian, which are :math:`J_a`, :math:`J_p`, :math:`A_a` and
+:math:`A_p`. These can be extracted by the attributes ``Ja``, ``Jp``,
+``Aa`` and ``Ap``, respectively, as shown below.
+
+.. code:: py
+
+   symbolic_Ja = jac.Ja
+   symbolic_Jp = jac.Jp
+   symbolic_Aa = jac.Aa
+   symbolic_Ap = jac.Ap
+
[…]
