+++ docs/examples.rst
@@ -1,0 +1,20 @@
+Mathematics
+===========
+
+.. toctree::
+   :maxdepth: 4
+
+   3R
+   4R4P
+   RSSRSSR
+
+
+Code
+====
+
+.. toctree::
+   :maxdepth: 4
+
+   3R_code
+   4R4P_code
+   RSSRSSR_code

[…]
