+++ src/acrod/jacobian.py
@@ -4,11 +4,24 @@ import itertools
 
 def check_robot_topology_matrix(M):
     """
-    check_robot_topology_matrix function performs several checks on the logical consistency of Robot-Topology matrix.
+    `check_robot_topology_matrix` function performs several checks on the logical consistency of Robot-Topology matrix.
 
     this function takes the robot-topology matrix as the input argument.
 
+    Parameters:
+    -----------
+
     :param M: robot-topology matrix (of size nxn, to be given in numpy.array format).
+    :type M: numpy.ndarray
+    
+    :raises ValueError: If the Robot-Topology matrix is not a 2D array.
+    :raises ValueError: If the Robot-Topology matrix is not a square matrix.
+    :raises ValueError: If any of the diagonal elements of the Robot-Topology matrix is not 9.
+    :raises ValueError: If the upper off-diagonal elements of the Robot-Topology matrix are other than the integers between 0 and 7.
+    :raises ValueError: If the lower off-diagonal elements of the Robot-Topology matrix are other than the integers 0 and 1.
+    :raises ValueError: If the Robot-Topology matrix has its lower off-diagonal element 1 but the corresponding upper off-diagonal element is 0.
+    :raises ValueError: If the Robot-Topology matrix has its lower off-diagonal element 1 but the corresponding upper off-diagonal element other than the integers 1 and 2.
+
     """
     s = M.shape
     if len(s) != 2:
@@ -35,13 +48,16 @@ def check_robot_topology_matrix(M):
 
 def all_joints_connected_to_the_link(M,linknumber):
     """
-    all_joints_connected_to_the_link function gives a list of all the joints connected to a given link for a given robot-topology matrix.
+    `all_joints_connected_to_the_link` function gives a list of all the joints connected to a given link for a given robot-topology matrix.
 
     this function takes the robot-topology matrix and the link number as input arguments.
 
     :param M: robot-topology matrix (of size nxn, to be given in numpy.array format).
+    :type M: numpy.ndarray
     :param linknumber: link number (it is to be indexed from 0, not from 1).
+    :type linknumber: int
     :return: a list of size n-1 items specifying the type of joint with which the given link is connected to every other joint. The order of joints corresponds to the ascending order of the link numbers.
+    :rtype: list
     """
     n=len(M)
     k=list(range(0,linknumber))+list(range(linknumber+1,n))
@@ -54,13 +70,16 @@ def all_joints_connected_to_the_link(M,linknumber):
 
 def all_links_connected_to_the_link(M,linknumber):
     """
-    all_links_connected_to_the_link function gives a list of all the links connected to a given link for a given robot-topology matrix.
+    `all_links_connected_to_the_link` function gives a list of all the links connected to a given link for a given robot-topology matrix.
 
     this function takes the robot-topology matrix and the link number as input arguments.
 
     :param M: robot-topology matrix (of size nxn, to be given in numpy.array format).
+    :type M: numpy.ndarray
     :param linknumber: link number (it is to be indexed from 0, not from 1).
+    :type linknumber: int
     :return: a list of items specifying the indices of links with which the given link is connected, in ascending order of the link numbers.
+    :rtype: list
     """
     k = all_joints_connected_to_the_link(M,linknumber)
     k2 = k[:linknumber]+[9]+k[linknumber:]
@@ -70,12 +89,14 @@ def all_links_connected_to_the_link(M,linknumber):
 # Function for getting all possible paths
 def all_paths(M):
     """
-    all_paths function gives a list of all paths connecting the base link and the end-effector link for a given robot-topology matrix.
+    `all_paths` function gives a list of all paths connecting the base link and the end-effector link for a given robot-topology matrix.
 
     this function takes the robot-topology matrix as input argument.
 
[…]
