+++ src/acrod/jacobian.py
@@ -329,9 +329,9 @@ class Jacobian(object):
         self.indices_P_tilde_omega = indices_P_tilde_omega
     def execute_equations(self):
         """
-        `execute_equations` is a method that takes no input arguments but takes the attributes `M`, `P`, `type`, `indices_P_tilde` and `indices_P_tilde_omega` of the class, and updates the attributes `superfluous_dof_information`, `Ja`, `Jp`, `Aa`, `Ap`, `parameters`, `parameters_symbolic`, `endeffector_variables_symbolic`, `Ja_func`, `Jp_func`, `Aa_func`, `Ap_func`, `active_joint_velocities`, `active_joint_velocities_symbolic`, `passive_joint_velocities` and `passive_joint_velocities_symbolic`. For serial manipulator, passive joints do not come into picture, and hence the attributes `Jp`, `Aa`, `Ap`, `Jp_func`, `Aa_func`, `Ap_func`, `passive_joint_velocities` and `passive_joint_velocities_symbolic` would remain with `None` value. Likewise, for manipulators without superfluous DOF, the attribute `superfluous_dof_information` would remain with `None` value. Additionally, it returns the $J_a$ function `Ja_func`, the $J_p$ function `Jp_func`, the $A_a$ function `Aa_func`, the $A_p$ function `Ap_func`, the list of active joint velocities (not symbolic) `active_jointvelocities`, the list of passive joint velocities (not symbolic) `passive_jointvelocities`, and the list of decision variables (not symbolic) `decision_variables_str`.
+        `execute_equations` is a method that takes no input arguments but takes the attributes `M`, `P`, `type`, `indices_P_tilde` and `indices_P_tilde_omega` of the class, and updates the attributes `superfluous_dof_information`, `Ja`, `Jp`, `Aa`, `Ap`, `parameters`, `parameters_symbolic`, `endeffector_variables_symbolic`, `Ja_func`, `Jp_func`, `Aa_func`, `Ap_func`, `active_joint_velocities`, `active_joint_velocities_symbolic`, `passive_joint_velocities` and `passive_joint_velocities_symbolic`. For serial manipulator, passive joints do not come into picture, and hence the attributes `Jp`, `Aa`, `Ap`, `Jp_func`, `Aa_func`, `Ap_func`, `passive_joint_velocities` and `passive_joint_velocities_symbolic` would remain with `None` value. Likewise, for manipulators without superfluous DOF, the attribute `superfluous_dof_information` would remain with `None` value. Additionally, it returns the :math:`J_a` function `Ja_func`, the :math:`J_p` function `Jp_func`, the :math:`A_a` function `Aa_func`, the :math:`A_p` function `Ap_func`, the list of active joint velocities (not symbolic) `active_jointvelocities`, the list of passive joint velocities (not symbolic) `passive_jointvelocities`, and the list of decision variables (not symbolic) `decision_variables_str`.
         
-        :return: The $J_a$ function `Ja_func`, the $J_p$ function `Jp_func`, the $A_a$ function `Aa_func`, the $A_p$ function `Ap_func`, the list of active joint velocities (not symbolic) `active_jointvelocities`, the list of passive joint velocities (not symbolic) `passive_jointvelocities`, and the list of decision variables (not symbolic) `decision_variables_str`.
+        :return: The :math:`J_a` function `Ja_func`, the :math:`J_p` function `Jp_func`, the :math:`A_a` function `Aa_func`, the :math:`A_p` function `Ap_func`, the list of active joint velocities (not symbolic) `active_jointvelocities`, the list of passive joint velocities (not symbolic) `passive_jointvelocities`, and the list of decision variables (not symbolic) `decision_variables_str`.
         :rtype: tuple
         """
         M = self.M
@@ -495,7 +495,7 @@ class Jacobian(object):
         """
[…]
