### 2023-12-31 suneeshjacob/ACRoD#11 @suneeshjacob
I made the documentation and it can be found [here](https://acrod.readthedocs.io/en/latest). I also added the link in the `README.md` file of the repository.
