### 2023-12-31 #5927 @suneeshjacob
> The repository is starting to look much better, especially the examples. However, some more work is needed. I opened two issues regarding the work which needs to be completed.
>
> 1. [Convert Examples to Notebooks suneeshjacob/ACRoD#10](https://github.com/suneeshjacob/ACRoD/issues/10) Examples are currently in a markdown file which it is not possible to directly run. I would like to see these moved to a Jupyter notebook or some other form in which they can be run directly.
> 2. [Package Documentation suneeshjacob/ACRoD#11](https://github.com/suneeshjacob/ACRoD/issues/11) There is currently not a centralized location to find documentation of all the package functions.
>
> Once these issues are solved, I believe this submission will be ready for publication.

@CameronDevine i resolved both the issues
