### 2024-01-11 #5927 @gfadini
Hi, thanks for the complete reply to my previous message.
I've seen the modifications to the code, I think they answer to some of the issues previously raised.
Here you find further discussion about some of your comments and some further questions.

## Previous comments
1. (i) If the library goal is to ease robot design I was expecting that at least some simple jacobian related metrics could be easily retrieved.
	So I find the example you added in the main `README` quite useful.
   (ii) I think I have been unclear on this point. I was not saying that the closed-loop jacobians are automatically computed by the existing libraries.
	Rather my point was that several libraries can be readily extended to take kinematic loops into account.
	As an example, the paper I have cited proves exactly such an approach.
	(iii) The usability and user-friendliness of the Robot-topology matrix are still debatable.
	A custom formulation (which may still be used internally by your library) may be rather limiting.
	Moreover:
	- the user needs to fill a matrix, potentially of huge size, which is an extremely error-prone process.
	- the information about the fixed design parameters (e.g. the link lengths you may not want to optimize) is not stored anywhere and needs to be specified later.
	URDF or SRDF would provide a starting point to avoid part of the process of building the topology matrix from scratch.
	They would generally hold the same level of information (and possibly also geometric properties).

2. I think these exact points should be much more clearly stated in the paper describing the library goals.

## On the library
7. From the mathematical point of view, the computation of sensitivities of the end effector has little to do with the fact of the joint being actuated.
	Selecting the active and passive joints has more to do with the choice of dependent DoFs rather than the actuation.
	Differently from serial robots, in closed loops, the independent coordinates need to be selected.
	From this choice, you can build	different Robot-Topology matrices. This step is potentially arbitrary and needs user specification (hence not automatic).

## Testing
2. The concept of joint type should not be confused with the actuation system.
	For instance one of the joints in [3..7] could represent a functional subgroup that globally acts with the corresponding DoFs.
	So the mathematical expressions to obtain the jacobians should be extended also to these joints, regardless of the actuation.
4. ~~Can you convert the examples .md files into actual runnable scripts?~~

## Formulation of the Jacobian
- I've noticed that the jacobian is computed with the end effector point as an input, I find this rather unusual.
	It seems weird that the expression does not depend on joint configuration $q$, say $J(q)$. Can you comment on this?
- Based on the examples so far the library seems to provide symbolic expressions.
	Then the numeric evaluation is obtained by specifying the configuration parameters.
	I think these details are not well documented yet.
	For instance, in the example what is exactly being expressed in the list `configuration_parameters = [3, 4, 2, 1, 4, 2]`?

## Additional comments on the paper
- line 10 should read the joint velocities (not only the actuated ones)
- line 32 "that can be used for any non-redundant manipulator" -> not clear what you mean here
