### 2024-01-17 a52363a
Update urdf_to_robottopologymatrix.py
