### 2024-01-17 b5fc8b1
Update urdf_to_robottopologymatrix.py
