### 2024-01-17 e9f1c38
Update urdf_to_robottopologymatrix.py
