+++ src/acrod/urdf_to_robottopologymatrix.py
@@ -13,14 +13,14 @@ def get_joints_from_urdf(root):
                     if child2.tag == 'child':
                         child_link = child2.attrib['link']
                 joints_info.append((child.attrib['name'], child.attrib['type'], parent_link, child_link,))
-    return sorted(joints_info, key=lambda x:x[0])
+    return numpy.unique(sorted(joints_info, key=lambda x:x[0]))
 
 def get_links_from_urdf(root):
     links_info = []
     for child in root:
         if child.tag == 'link':
             links_info.append(child.attrib['name'])
-    return sorted(links_info)
+    return numpy.unique(sorted(links_info))
 
 import xml.etree.ElementTree as ET
 
