### 2024-01-17 0101dfb
Update urdf_to_robottopologymatrix.py
