### 2024-01-17 eff22e2
Update urdf_to_robottopologymatrix.py
