### 2024-01-17 5bda2f1
Update urdf_to_robottopologymatrix.py
