### 2024-01-17 b461977
Update urdf_to_robottopologymatrix.py
