### 2024-01-17 6c21419
Update urdf_to_robottopologymatrix.py
