### 2024-01-17 1125b5a
Update urdf_to_robottopologymatrix.py
