### 2024-01-17 a89551f
Update urdf_to_robottopologymatrix.py
