### 2024-01-17 dba069d
Update urdf_to_robottopologymatrix.py
