### 2024-01-17 4da0514
Update urdf_to_robottopologymatrix.py
