### 2024-01-17 aa7f004
Update urdf_to_robottopologymatrix.py
