### 2024-01-17 5492f29
Update urdf_to_robottopologymatrix.py
