### 2024-01-17 a5ccb4d
Update urdf_to_robottopologymatrix.py
