### 2024-01-17 a714a7c
Update urdf_to_robottopologymatrix.py
