### 2024-01-17 12032c5
Update urdf_to_robottopologymatrix.py
