### 2024-01-17 5a35cbc
Update urdf_to_robottopologymatrix.py
