### 2024-01-17 80bfb8a
Update urdf_to_robottopologymatrix.py
