### 2024-01-21 #5927 @JHartzer
I believe the authors have addressed my main concerns and have completed all but one of the issues submitted to their repository. The one open task is https://github.com/suneeshjacob/ACRoD/issues/3#issue-1934220317 and it appears to be complete, just not closed. As far as what I have brought up this completes my review.

I do believe that @gfadini has brought up some good points as well that should be addressed.
