### 2024-01-31 3c52ff2
Update urdf_to_robottopologymatrix.py
