### 2024-02-02 ab88dfc
Update urdf_to_robottopologymatrix.py
