### 2024-02-02 b2b3a0c
Update urdf_to_robottopologymatrix.py
