### 2024-02-02 c08cb3c
Update urdf_to_robottopologymatrix.py
