### 2024-02-02 2592e9a
Update urdf_to_robottopologymatrix.py
