+++ docs/urdf_to_robot-topology_matrix.rst
@@ -1,0 +1,26 @@
+URDF support
+============
+
+URDF to Robot-topology Matrix
+-----------------------------
+
+Support for getting the robot's topology from a URDF file is provided in this section.
+
+
+
+-  For contribution to the software:
+
+   -  In order to contribute to the software, please consider using the
+      `pull request
+      feature <https://github.com/suneeshjacob/ACRoD/pulls>`__ of
+      GitHub.
+
+-  For reporting issues with the software:
+
+   -  For reporting issues or problems, please use
+      `issues <https://github.com/suneeshjacob/ACRoD/issues>`__.
+
+-  For support:
+
+   -  For any further support (including installation, usage, etc.),
+      feel free to contact via suneeshjacob-at-gmail-dot-com.
