### 2024-02-03 #5927 @suneeshjacob
> Hi, thanks for the complete reply to my previous message. I've seen the modifications to the code, I think they answer to some of the issues previously raised. Here you find further discussion about some of your comments and some further questions.
>
> ## Previous comments
> 1. (i) If the library goal is to ease robot design I was expecting that at least some simple jacobian related metrics could be easily retrieved.
>    So I find the example you added in the main `README` quite useful.
>    (ii) I think I have been unclear on this point. I was not saying that the closed-loop jacobians are automatically computed by the existing libraries.
>    Rather my point was that several libraries can be readily extended to take kinematic loops into account.
>    As an example, the paper I have cited proves exactly such an approach.
>    (iii) The usability and user-friendliness of the Robot-topology matrix are still debatable.
>    A custom formulation (which may still be used internally by your library) may be rather limiting.
>    Moreover:
>
>    * the user needs to fill a matrix, potentially of huge size, which is an extremely error-prone process.
>    * the information about the fixed design parameters (e.g. the link lengths you may not want to optimize) is not stored anywhere and needs to be specified later.
>      URDF or SRDF would provide a starting point to avoid part of the process of building the topology matrix from scratch.
>      They would generally hold the same level of information (and possibly also geometric properties).
> 2. I think these exact points should be much more clearly stated in the paper describing the library goals.
>
> ## On the library
> 7. From the mathematical point of view, the computation of sensitivities of the end effector has little to do with the fact of the joint being actuated.
>    Selecting the active and passive joints has more to do with the choice of dependent DoFs rather than the actuation.
>    Differently from serial robots, in closed loops, the independent coordinates need to be selected.
>    From this choice, you can build	different Robot-Topology matrices. This step is potentially arbitrary and needs user specification (hence not automatic).
>
> ## Testing
> 2. The concept of joint type should not be confused with the actuation system.
>    For instance one of the joints in [3..7] could represent a functional subgroup that globally acts with the corresponding DoFs.
>    So the mathematical expressions to obtain the jacobians should be extended also to these joints, regardless of the actuation.
> 3. ~Can you convert the examples .md files into actual runnable scripts?~
>
> ## Formulation of the Jacobian
> * I've noticed that the jacobian is computed with the end effector point as an input, I find this rather unusual.
>   It seems weird that the expression does not depend on joint configuration q, say J(q). Can you comment on this?
> * Based on the examples so far the library seems to provide symbolic expressions.
>   Then the numeric evaluation is obtained by specifying the configuration parameters.
>   I think these details are not well documented yet.
>   For instance, in the example what is exactly being expressed in the list `configuration_parameters = [3, 4, 2, 1, 4, 2]`?
>
> ## Additional comments on the paper
> * line 10 should read the joint velocities (not only the actuated ones)
> * line 32 "that can be used for any non-redundant manipulator" -> not clear what you mean here

@gfadini Hi, thank you for going into depth of the technical aspects of the manuscript and providing additional comments. I am posting the responses to your comments below.

1.
(iii) Yes the usability of robot-topology matrix might be limiting in the sense that it is not a widely used notation (or slightly deviates from the adjacency matrix which is a widely used notation), however I think a URDF file as an alternative would have similar (or even more) limitations. Firstly, if filling a matrix to describe a robot's topology is error-prone, then writing a URDF file to describe the robot's topology would also be error-prone, unless a URDF file or a matrix is already available.
However, I made a simple script file (perhaps still beta) to convert a URDF file into its corresponding robot-topology matrix. It works well with two example files provided in the repository.

2. I added a sentence at the end of the 'Statement of need' portion of the paper, highlighting its exact purpose.

On the library:

7. I worked out mathematical analysis of a simple planar 5-bar (2-DOF) closed-loop manipulator of all revolute joints [here in this link](https://drive.google.com/file/d/1HdxroQAxp8kibwn6SAfmYcVnLGZoVob8/view?usp=drive_link), which shows that the Jacobian (including its singular values and its condition number) would depend on which joints are considered as active joints. Hence for closed-loop chains, the independent coordinates need to be specified in order to determine the Jacobian. Hence, the lower off-diagonal part of the robot-topology matrix is required to be specified, if the designer wants to have a deterministic Jacobian.

Testing:

2. I don't understand what you mean by a joint representing 'a functional subgroup that globally acts with the corresponding DoFs.' The point I was trying to make is that considering practical usage, only prismatic and revolute type joints are considered as possible actuators in this framework. One may mathematically make some formulations by assuming a universal joint as an actuator and find some hypothetical Jacobian matrix at a given point, but I don't think it would have much practical utility. And in order to accommodate such cases, a lot of changes in the code may have to be made, with very less practical utility, apparently. Even if somehow it were made, that wouldn't be complete because one may also think of a 2 DOF case where one revolute motion from one universal joint and another revolute motion from aother universal joint would be the independent actuating joint velocities and the rest are passive. In such case, the robot-topology matrix would be insufficient to describe only one of the two revolute motions of a universal joint as an actuating joint velocity. I have not seen universal joints, cylindrical joints, helical joints, etc., as actuators. Even though there may be some studies on cylindrical actuators, it seems that such spherical actuation behind the scenes reduces to three revolute joints having three mutually perpendicular concurrent axes in 3D space. Hence, if at all such cases arise, I think it is better to consider those joints as combinations of revolute and prismatic joints and consider some additional constraints in the optimisation formulation for dimensional synthesis. For example, a universal joint can be considered to be a set of two links with two revolute joints, where one of the links has the link-length as zero and the axes of the two revolute joints are intersecting and mutually perpendicular to each other, which would be the constraints to be supplied in the optimisation problem (of some performance indices). Likewise, a cylindrical joint can be considered as a combination of a revolute joint and a prismatic joint, with some constraints, etc.

Formulation of the Jacobian:

1.
1a) Yes, the Jacobian is computed here with the end-effector as one of the inputs. This is a limitation of the current repository that it is valid only for a single task point. In future it is planned to implement the same for multiple task points that could span the entire workspace.
1b) The expression depends on $q$, but not explicitly. For example, the Jacobian of a 2R planar robot is
[[-a_y+r_12y, -a_y+r_23y], [a_x-r_12x, a_x-r_23x], [1, 1]]. The typical Jacobian for a 2R planar robot in terms of $q$ would be [[-l_1 sin(theta_1)-l_2 sin(theta_1+theta_2), [-l_2 sin(theta_1+theta_2)]], [l_1 cos(theta_1)+l_2 cos(theta_1+theta_2), l_2 cos(theta_1+theta_2)], [1, 1]], where $q$ = [theta_1, theta_2]. These expressions are equivalent. For example, the quantity -l_1 sin(theta_1)-l_2 sin(theta_1+theta_2) is the same as -a_y+r_12y, but expressed in different parameters. The former is written with positions of joints as variable parameters whilst the latter is written in terms of link-lengths and joint angles as variable parameters. From the point of dimensional synthesis, both would be equivalent because the objective would be to find the optimal link-lengths which can be found either by optimising the condition number of the former expression or by optimising the condition number of the latter expression and calculating the link-lengths accordingly. The expressions in the former case are easy to compute (as they do not involve sine and cosine expressions), however a post-processing to extract link-lengths from the optimal solution would be required. In the latter case, the expressions are more complicated than that of the former case, however the optimal solutions would directly contain the link-lengths without any post-processing. So, each case is kind of equivalent. This repository uses the former way (for its easy accommodation of specifying the end-effector point), hence the Jacobian is not an explicit function of $q$, although implicitly it indeed is a function of $q$. And from the point of view of dimensional synthesis, it is expected to give the same result in both the cases.

2. The meaning of `configuration_parameters = [3, 4, 2, 1, 4, 2]` is documented right above the blockcode. Those six numbers are dimensional parameters, which are the numerical values of the expressions in `robot_dimensional_parameters` in the same order. They basically involve the positions of joints, the orientations of axes, etc. In that example, the robot_dimensional_parameters are [r_12x, r_12y, r_23x, r_23y, r_34x, r_34y], where r_12 = [r_12x, r_12y], where r_12x and r_12y are the x and y components of the position vector of the location of the joint connecting links 1 and 2, and similary for r_23 and r_34. And since in the sample configuration r_12 = [3, 4], r_23 = [2, 1] and r_34 = [4,2], the configuration parameters happened to be [3, 4, 2, 1, 4, 2].

Additional comments on the paper:

1. Why should it not be the actuated joints? If other joints are considered then that wouldn't be the typical Jacobian in the context of Robotics. In the context of mathematics, Jacobian is a mapping of the first-order partial derivatives that transform the infinitesimals from one set of coordinates to another set of coordinates. However, Jacobian in the context of Robotics is the mapping of actuating joint velocities and end-effector velocities. In case of serial manipulators, all the joints are actuated and hence it would be enough to say that it is a mapping of joint velocities and end-effector velocities. However, in parallel manipulators the Jacobian would depend on which joints are considered active joints (and which are considerd passive). Hence, I think it should be actuating joint velocities. If not, then can you explain why not?

2. It says this Jacobian formulation module can be used for any non-redundant manipulator, i.e., any manipulator that has the number of degrees of freedom of the end-effector link to be the same as the number of actuating joints.
