### 2024-02-04 65123db
Made some changes to automatic tests
