### 2024-02-04 suneeshjacob/ACRoD#3 @suneeshjacob
Hi, I now made changes in the tests to let it print - whenever there is a mismatch in a numpy matrix - the row and column indices where the first mismatch is found, so as to make it easier for developers to debug any mismatch between expected and computed values. Thanks for your suggestion.

With this, I am closing this issue.
