### 2024-02-04 #5927 @suneeshjacob
> I believe the authors have addressed my main concerns and have completed all but one of the issues submitted to their repository. The one open task is [suneeshjacob/ACRoD#3 (comment)](https://github.com/suneeshjacob/ACRoD/issues/3#issue-1934220317) and it appears to be complete, just not closed. As far as what I have brought up this completes my review.
>
> I do believe that @gfadini has brought up some good points as well that should be addressed.

@JHartzer I added a feature to print the row and column indices of first elemental mismatch detected in the pytests, and I now closed the issue.
