### 2024-02-15 #5927 @gfadini
@suneeshjacob thanks for your reply I here you find some additional comments, please tell me if anything is unclear:

1. (iii) yes `urdf` files are easily available, (so you are not even writing it) that's the whole point of my comment. Even if I could not test it, I think that the addition of the script is going in the right direction. No need to reinvent the wheel :)
2. Can you please also modify the `README` so it reflects the changes?
7. Yes, exactly. so you agree with me: there is still some choice that the user has to make, hence the process is not fully automatic.

Testing:
2. Imagine you want to study a multi-DOF joint, what I referred to as "functional subgroup" is what you are suggesting in the reply (an ensemble of actuators that globally allows a certain movement). Now I do not agree with your argument, the considerations to be made here are mathematical and not practical because (i) as I remarked in a previous comment, the idea of dependent/independent joints (mathematical concept) is not to be confounded with the one of active/passive joint (which is related to actuators and the practical realization) (ii) the fact that a given motion can be obtained with a given composition of joints (spherical as RRR) does not mean that such composition is the right mathematical model (e.g. you can still encounter in problems such as gimbal lock), hence having general mathematical joint models is still valuable even if, in practice, there may not exist actuators to produce those motions directly.

Formulation of the jacobian:
1.a This is still rather unclear: how do you enforce constraints in your jacobian formulation? If you use your method on the RR toy problem you come up with 6 variables $a_x, a_y, a_12_x, a_12_y, a_23_x, a_23_y$ while the "usual" representation needs 4 $l_1, l_2, \theta_1, theta_2$, how are those two linked? My intuition is that there should be 2 hidden constraints to keep the two formulations consistent. How do you enforce them? Can you comment on this?
2. Ok, maybe you can consider adding some more information in the doc about it. As these are automatically generated strings, they may become rather ugly/unclear to read.

Additional comments on the paper:
1. I would rather use "independent" instead of "active" or "actuated" in this context and also in other paper parts. As I pointed out in the previous comment they are different concepts. For example in a floating base robot all joints are independent (no loop closure) but not all joints are actuated (the base cannot be freely moved in space). I think that "passive joint" is generally used as a synonym of "non-actuated" and not "dependent". I suggest you to propagate this change in the doc and paper.
2. that's clear now 👍🏻

Overall the paper and the library quality is increased.
However @Kevin-Mattheus-Moerman before recommending the submission for publication, I'd still wait a reply on the method (in particular 1.a) from @suneeshjacob. Moreover some modifications about naming are still recommended (Point 1. of additional comments).
