### 2024-02-15 64a0356
Update README.md
Added some lines regarding the absence of optimisation scripts based on JOSS reviewer `gfadini`'s comment.
