### 2024-02-15 #5927 @suneeshjacob
> @suneeshjacob thanks for your reply I here you find some additional comments, please tell me if anything is unclear:
>
> 1. (iii) yes `urdf` files are easily available, (so you are not even writing it) that's the whole point of my comment. Even if I could not test it, I think that the addition of the script is going in the right direction. No need to reinvent the wheel :)
> 2. Can you please also modify the `README` so it reflects the changes?
> 3. Yes, exactly. so you agree with me: there is still some choice that the user has to make, hence the process is not fully automatic.
>
> Testing: 2. Imagine you want to study a multi-DOF joint, what I referred to as "functional subgroup" is what you are suggesting in the reply (an ensemble of actuators that globally allows a certain movement). Now I do not agree with your argument, the considerations to be made here are mathematical and not practical because (i) as I remarked in a previous comment, the idea of dependent/independent joints (mathematical concept) is not to be confounded with the one of active/passive joint (which is related to actuators and the practical realization) (ii) the fact that a given motion can be obtained with a given composition of joints (spherical as RRR) does not mean that such composition is the right mathematical model (e.g. you can still encounter in problems such as gimbal lock), hence having general mathematical joint models is still valuable even if, in practice, there may not exist actuators to produce those motions directly.
>
> Formulation of the jacobian: 1.a This is still rather unclear: how do you enforce constraints in your jacobian formulation? If you use your method on the RR toy problem you come up with 6 variables ax,ay,a12x,a12y,a23x,a23y while the "usual" representation needs 4 l1,l2,θ1,θ2, how are those two linked? My intuition is that there should be 2 hidden constraints to keep the two formulations consistent. How do you enforce them? Can you comment on this? 2. Ok, maybe you can consider adding some more information in the doc about it. As these are automatically generated strings, they may become rather ugly/unclear to read.
>
> Additional comments on the paper:
>
> 1. I would rather use "independent" instead of "active" or "actuated" in this context and also in other paper parts. As I pointed out in the previous comment they are different concepts. For example in a floating base robot all joints are independent (no loop closure) but not all joints are actuated (the base cannot be freely moved in space). I think that "passive joint" is generally used as a synonym of "non-actuated" and not "dependent". I suggest you to propagate this change in the doc and paper.
> 2. that's clear now 👍🏻
>
> Overall the paper and the library quality is increased. However @Kevin-Mattheus-Moerman before recommending the submission for publication, I'd still wait a reply on the method (in particular 1.a) from @suneeshjacob. Moreover some modifications about naming are still recommended (Point 1. of additional comments).

@gfadini Hi, thanks for your response. Here is my response to it:

2. I added a similar statement at the end of 'Statement of need' in the README.md file.
3. I have both agreement and disagreement at certain aspects of your argument. I disagree with your statements “From the mathematical point of view, the computation of sensitivities of the end effector has little to do with the fact of the joint being actuated” and “Selecting the active and passive joints has more to do with the choice of dependent DoFs rather than the actuation.” And because of this dependency of singular values upon which joints are considered as actuators, the non-automation of off-diagonal elements is attributed not to a left-out work in the code that needs to be added but rather to its non-deterministic nature of Jacobian under the absence of specification of which joints are actuated. That was the point i was trying to make. And yes there is some choice the user has to make, in which case the preparation of robot-topology matrix is not automatically done for a mere adjacency matrix. I have agreement with you on that part. But that doesn’t negate the meaning of automation in the repository in its context, because the term automation claimed in the repository is about the automation from the input (which already includes the user’s choice of off-diagonal elements in the input matrix) to the production of functions required to compute Jacobian.

Testing:
(i) and (ii) i do understand that there is a difference between active/passive classification and dependent/independent classification. However, what i don’t understand is how the dependent/independent classification is relevant w.r.t. your suggestion. Your suggestion is to facilitate the code to include cylindrical, spherical, universal, helical and planar joints for serial manipulators. For example, if there is a 4-link serial robot with joints R, U, S (let us say). Even though actuation of joints such as U and S doesn’t have much practical usage, in theory one should mathematically be able to formulate Jacobian. Am i understanding your question correctly? If yes, then my response to it is, even in such a case of mathematical formulation of Jacobian by means of considering some set of ‘independent’ joint velocities, that would require which independent joint velocities in order to determine that Jacobian. Else, it would simply be as good as a set of implicit equations, from which a deterministic Jacobian cannot be extracted. And in order to numerically separate them into explicit equations, one needs to consider the independent set of joint velocities, and the Jacobian depends on which set of joint velocities are chosen as independent. Hence, the specification of independent joint velocities would be necessary. And as far as robot-topology matrix is concerned, it would not be able to specify partial degrees of freedom of a joint as independent joint velocities, and hence, in order to facilitate all such possibilities, almost the whole structure has to be changed, with limited utility (limited utility in the sense that they are practically not being used much). Hence I was suggesting that instead of changing almost the whole structure of the code to handle such special cases, it would be better that this repository has this limitation of actuators being among only prismatic and revolute joints (in which case the Jacobian would be a mapping of end-effector velocities with only prismatic and revolute joints which are actuators, in which case a serial robot with joints other than these two joints would fall outside of the scope of the repository), and that to suggest that whenever other joints are required to be considered as actuators (or independent joint velocities) some appropriate additional constraints may be incorporated into the optimisation process for dimensional synthesis. And regarding the gimbal lock issue that you raised in considering spherical as RRR, as long as additional constraints can be put such that the axes of all the three revolute joints must be mutually perpendicular in all configurations (or any such appropriate set of constraints), it is quite possible to represent angular velocity along any arbitrary axis as the sum of components of individual velocities around those mutually perpendicular axes, thereby eliminating gimbal lock case (in such case, designing the controllers with such constraints might be complicated, but for the sake of dimensional synthesis it would be sufficient to replace a spherical joint with three revolute joints, apparently). And hence, any motion that a spherical joint could produce can be made to be produced by the combination of three revolute joints, apparently.

Formulation of the Jacobian:
1.a You have rightly pointed out that the usual representation has 4 independent parameters whilst the location-and-orientation-based representation has 6 parameters and that there must be two (not very hidden) constraints. Well, in your case of “usual” representation, you assumed that the base of the robot is at the origin from which the links would start. That is a condition that restricts both x and y coordinates of the location of the first joint (theta_1). And in the other case, it instead considers the parameters determining the position of the first joint (i.e., the location represented by $r\_{(1,2)x}$ and $r_{(1,2)y}$) as variables, and hence two more parameters happened to be there in the set of input parameters. In the example that i added in the README.md file, i used (0,0) as the location of that joint (by the line `base_reference_point = [0,0]` in the code), however, for parallel manipulators, there could be various joints attached to the end-effector, and either of them could be considered to be at the origin (and for that matter, any joint - whether attached to the base link or not - can be considered to be at the origin). So, I would leave it to the designer to choose which joint to be considered to be at the origin. Nevertheless, this falls under the purview of defining optimisation problem which does not fall under the formulation of Jacobian per se. Even if such a constraint had not been used, the algorithm would still choose a valid base point at the end of the optimisation process (although one may debate on the practical usage of it).

2. They are not automatically generated strings, but rather entered by me manually. And in case you haven't already noticed, i added [API documentation](https://acrod.readthedocs.io/en/latest/index.html) for the repository using `readthedocs` based on the suggestion by [CameronDevine](https://github.com/CameronDevine), which includes this as well. And last time i checked, they were pretty clear to me. You can check it out. If you find them unclear in this readthedocs documentation, please let me know.

Additional comments on the paper:
1. Sorry but i have to disagree again here. Your suggestion of the usage of the word ‘independent’ gives a very different meaning. It is not the case that Jacobian of a robot is the mapping of end-effector velocities and any set of independent joint velocities. As i mentioned earlier in the worked-out document with the example of 5-bar closed-loop mechanism, there are multiple sets of independent coordinates (or generalised coordinates) possible, that can be considered to determine the velocity of the end-effector. In that worked-out example, let’s suppose $\theta_2$ and $\theta_3$ are considered as actuated joints. $\dot{\theta}_2$ and $\dot{\theta}_4$ can be considered as independent set of joint velocities in the sense that having known these two values would be sufficient to determine other joint velocities. Now if i say that ‘Jacobian is a mapping of end-effector velocities and independent joint velocities’ or if i say that ‘Jacobian is a mapping of end-effector velocities and joint velocities’ without mentioning the term ‘actuated’, that would imply that a mapping that maps end-effector velocities and the set $\dot{\theta}_2$ and $\dot{\theta}_4$ (for which a mapping matrix is possible) would be a valid Jacobian for the robot with $\theta_2$ and $\theta_3$ as actuating joints. That would be technically not true. The Jacobian of a robot (as defined in [1], [2] and [3]) would be a matrix that maps end-effector velocities and actuating joint velocities. And as far as floating base robots are considered, this repository is currently supported only for robots with fixed base link, and hence mobile robots and floating base robots would fall outside the scope of this repository, although they may be added in future releases.
[1] Jean-Pierre Merlet. Jacobian, manipulability, condition number and accuracy of parallelrobots. ISRR, Oct 2005, San Francisco, 2005.
[2] Liu, Haitao, Tian Huang, and Derek G. Chetwynd. "A method to formulate a dimensionally homogeneous Jacobian of parallel manipulators." IEEE Transactions on Robotics 27.1 (2010): 150-156.
[3] Hoevenaars, A. G. L., S. Krut, and J. L. Herder. "Jacobian-based natural frequency analysis of parallel manipulators." Mechanism and Machine Theory 148 (2020): 103775.
