### 2024-02-16 #5927 @gfadini
Thanks for making these points clear @suneeshjacob

> Even though actuation of joints such as U and S doesn’t have much practical usage, in theory one should mathematically be able to formulate Jacobian.

Yes, and this would allow to avoid the ill-posedness of the jacobian with these tailored mathematical formulations. It is a quite important point that you can keep in mind in future development as it may involve major modifications to the codebase.

> Nevertheless, this falls under the purview of defining optimisation problem which does not fall under the formulation of Jacobian per se. Even if such a constraint had not been used, the algorithm would still choose a valid base point at the end of the optimisation process (although one may debate on the practical usage of it).

I see, as you say, the practical applicability of the obtained jacobian alone may be questionable, but I get the logic behind it if the goal is just to obtain intermediate representations to then use in further optimization.

I think Point 3. and the last one are tied. "Actuated" generally means that on the joint a control action is applied to govern its motion. The fact that the independent joints coordinates are chosen where the "actuated" joints are located is a possible choice to get a deterministic jacobian, and it is generally reasonable. So in your work this equivalence is correct. However, from a more general theoretical point of view, the naming "active" or "actuated" is misleading (even if common) because the jacobian is only related to the mechanism kinematics. I think another option would be "respect the generalized joint coordinates" as common literature. This is not a crucial modification as I think it is sufficiently clear what you are trying to do.

@Kevin-Mattheus-Moerman everything is fine by me now, you can proceed for publication 👍🏻
