+++ paper/paper.bib
@@ -46,6 +46,8 @@
   publisher = {arXiv},
   
   year = {2022},
+
+  doi = {10.48550/arXiv.2210.03327},
   
   copyright = {arXiv.org perpetual, non-exclusive license}
 }
@@ -55,6 +57,7 @@
       author={Akkarapakam Suneesh Jacob and Bhaskar Dasgupta},
       year={2022},
       archivePrefix={arXiv},
+      doi={10.48550/arXiv.2210.04446},
       primaryClass={cs.RO}
 }
 
@@ -204,19 +207,31 @@ doi = {10.1016/j.mechmachtheory.2007.08.001}
   doi = {10.7717/peerj-cs.103}
 }
 
-@inproceedings{carpentier:hal-01866228,
-  TITLE = {{The Pinocchio C++ library -- A fast and flexible implementation of rigid body dynamics algorithms and their analytical derivatives}},
-  AUTHOR = {Carpentier, Justin and Saurel, Guilhem and Buondonno, Gabriele and Mirabel, Joseph and Lamiraux, Florent and Stasse, Olivier and Mansard, Nicolas},
-  URL = {https://laas.hal.science/hal-01866228},
-  BOOKTITLE = {{SII 2019 - International Symposium on System Integrations}},
-  ADDRESS = {Paris, France},
-  HAL_LOCAL_REFERENCE = {Rapport LAAS n{\textdegree} 18288},
-  YEAR = {2019},
-  MONTH = Jan,
-  PDF = {https://laas.hal.science/hal-01866228v2/file/19-sii-pinocchio.pdf},
-  HAL_ID = {hal-01866228},
-  HAL_VERSION = {v2},
-}
+@INPROCEEDINGS{carpentier:hal-01866228,
+  author={Carpentier, Justin and Saurel, Guilhem and Buondonno, Gabriele and Mirabel, Joseph and Lamiraux, Florent and Stasse, Olivier and Mansard, Nicolas},
+  booktitle={2019 IEEE/SICE International Symposium on System Integration (SII)}, 
+  title={The Pinocchio C++ library : A fast and flexible implementation of rigid body dynamics algorithms and their analytical derivatives}, 
+  year={2019},
+  volume={},
+  number={},
+  pages={614-619},
+  keywords={Heuristic algorithms;Computational modeling;Kinematics;Robot kinematics;Libraries;Software algorithms},
+  doi={10.1109/SII.2019.8700380}}
+
+
+%@inproceedings{carpentier:hal-01866228,
+%  TITLE = {{The Pinocchio C++ library -- A fast and flexible implementation of rigid body dynamics algorithms and their analytical derivatives}},
+%  AUTHOR = {Carpentier, Justin and Saurel, Guilhem and Buondonno, Gabriele and Mirabel, Joseph and Lamiraux, Florent and Stasse, Olivier and Mansard, Nicolas},
+%  URL = {https://laas.hal.science/hal-01866228},
+%  BOOKTITLE = {{SII 2019 - International Symposium on System Integrations}},
+%  ADDRESS = {Paris, France},
+%  HAL_LOCAL_REFERENCE = {Rapport LAAS n{\textdegree} 18288},
+%  YEAR = {2019},
+%  MONTH = Jan,
+%  PDF = {https://laas.hal.science/hal-01866228v2/file/19-sii-pinocchio.pdf},
+%  HAL_ID = {hal-01866228},
+%  HAL_VERSION = {v2},
+%}
 
 @INPROCEEDINGS{10260660,
   author={Tola, Daniella and Corke, Peter},
