### 2024-03-14 80dc9df
Update paper.md
Changed `jacobian` to `Jacobian`
