### 2024-03-14 78dcb26
Changed `triped` to `tripedal`
