### 2024-03-14 c53da72
Revert "Changed `softwares` to `software`"
This reverts commit 32d20321ddeed8acdcf313c9fd0d72f41d42a72e.
