### 2024-03-14 #5927 @suneeshjacob
@Kevin-Mattheus-Moerman I did the following:

1) Replaced `jacobian` with `Jacobian` in all occurrences in the paper.
2) Changed the word `triped` to `tripedal` in all occurrences in the paper.
3) Changed the term `softwares` to `software packages` in all occurrences in the paper.
4) Verified the ORCIDs and affiliations of the authors.
5) Released a new version of the software: v1.2.0
6) Released this version (v1.2.0) in Zenodo: [10.5281/zenodo.10819242](https://doi.org/10.5281/zenodo.10819242)
7) Made sure the title, the authors, their ORCIDs, and the license are the same.

I think these appropriately address all the items in both the post-review checklist and the points you suggested. If there is anything required from my side, please let me know. Thanks!
