### 2024-03-14 #5927 @Kevin-Mattheus-Moerman
@suneeshjacob can you set the archive version (you can manually edit it) to `v1.2.0`, so to include the `v`?
