### 2024-03-21 fb95290
Update CITATION.cff
Added JOSS DOI
