### 2024-03-21 278f276
Update CITATION.cff
