### 2024-03-22 4f11f59
Bump readthedocs-sphinx-search from 0.1.1 to 0.3.2 in /docs
Bumps [readthedocs-sphinx-search](https://github.com/readthedocs/readthedocs-sphinx-search) from 0.1.1 to 0.3.2.
- [Changelog](https://github.com/readthedocs/readthedocs-sphinx-search/blob/main/CHANGELOG.rst)
- [Commits](https://github.com/readthedocs/readthedocs-sphinx-search/compare/0.1.1...0.3.2)

---
updated-dependencies:
- dependency-name: readthedocs-sphinx-search
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
