+++ paper/paper.bib
@@ -1,232 +1,0 @@
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-@INPROCEEDINGS{846382,  author={Kim, D. and Chung, W. and Youm, Y.},  booktitle={Proceedings 2000 ICRA. Millennium Conference. IEEE International Conference on Robotics and Automation. Symposia Proceedings (Cat. No.00CH37065)},   title={Analytic Jacobian of in-parallel manipulators},   year={2000},  volume={3},  number={},  pages={2376-2381 vol.3},  abstract={A general formulation to obtain an analytic Jacobian matrix for in-parallel manipulators has not been developed yet, although several authors addressed it. This paper suggests a consistent approach to obtain the instantaneous kinematic relation of in-parallel manipulators, by extending the results of screw theory to the forward and inverse instantaneous relations, and gives a formula that can obtain a reciprocal screw analytically without considering geometric relations.},  keywords={},  ISSN={1050-4729},  month={April}, doi={10.1109/robot.2000.846382},}
-
-
-
-
-
-@article{patel2015manipulator,
-  title={Manipulator performance measures-a comprehensive literature survey},
-  author={Patel, Sarosh and Sobh, Tarek},
-  journal={Journal of Intelligent \& Robotic Systems},
-  volume={77},
-  number={3},
-  pages={547--570},
-  year={2015},
-  publisher={Springer},
-  doi={10.1007/s10846-014-0024-y},
-}
-
-
-@misc{enumeration,
-  
-  author = {Jacob, Akkarapakam Suneesh and Dasgupta, Bhaskar and Datta, Rituparna},
-  
-  keywords = {Robotics (cs.RO), FOS: Computer and information sciences, FOS: Computer and information sciences},
-  
-  title = {Enumeration of spatial manipulators by using the concept of Adjacency Matrix},
-  
-  publisher = {arXiv},
-  
-  year = {2022},
-
-  doi = {10.48550/arXiv.2210.03327},
-  
-  copyright = {arXiv.org perpetual, non-exclusive license}
-}
-
-@misc{synthesis,
-      title={Dimensional synthesis of spatial manipulators for velocity and force transmission for operation around a specified task point}, 
-      author={Akkarapakam Suneesh Jacob and Bhaskar Dasgupta},
-      year={2022},
-      archivePrefix={arXiv},
-      doi={10.48550/arXiv.2210.04446},
-      primaryClass={cs.RO}
-}
-
-
-
-
-
-
-
-@ARTICLE{1220722,  author={Sung-Gaun Kim and Ryu, J.},  journal={IEEE Transactions on Robotics and Automation},   title={New dimensionally homogeneous Jacobian matrix formulation by three end-effector points for optimal design of parallel manipulators},   year={2003},  volume={19},  number={4},  pages={731-736}, doi={10.1109/tra.2003.814496}}
-
-
-@article{ALTUZARRA20061407,
-title = {Point-based Jacobian formulation for computational kinematics of manipulators},
-journal = {Mechanism and Machine Theory},
-volume = {41},
-number = {12},
-pages = {1407-1423},
-year = {2006},
-issn = {0094-114X},
-author = {O. Altuzarra and O. Salgado and V. Petuya and A. Hernández},
-keywords = {Velocity equation, Dimensionless Jacobian, Motion space, Parallel manipulators},
[…]
