+++ README.md
@@ -1,6 +1,6 @@
 # Automatic Computation for Robot Design (ACRoD) [![status](https://joss.theoj.org/papers/2c8c94adbd8240eadaa4f77c2057b14c/status.svg)](https://joss.theoj.org/papers/2c8c94adbd8240eadaa4f77c2057b14c)
 
-## Description (New Branch)
+## Description (For `next_version` branch)
 
 This branch is planned to publish Python scripts to automatically formulate Jacobian for multiple end-effector points, thereby removing the limitation of "single end-effector point" of ACRoD. After gathering successful scripts, this branch is planned to be merged with the main branch.
 
