+++ README.md
@@ -2,5 +2,5 @@
 
 ## Description (For `next_version` branch)
 
-This branch is planned to publish Python scripts to automatically formulate Jacobian for multiple end-effector points, thereby removing the limitation of "single end-effector point" of ACRoD. After gathering successful scripts, this branch is planned to be merged with the main branch.
+This branch is planned for publishing Python scripts to automatically formulate Jacobian for multiple end-effector points, thereby removing the limitation of "single end-effector point" of ACRoD. After gathering successful scripts, this branch is planned to be merged with the main branch.
 
