### 2024-03-23 fd2b369
Update README.md
Removed the JOSS tag that was mistakenly retained
